% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeBuildingFloorArea.R
\name{calcCeBuildingFloorArea}
\alias{calcCeBuildingFloorArea}
\title{Calculates relative floor area of SF, MF and NR buildings by structure type.}
\usage{
calcCeBuildingFloorArea(subtype = NULL)
}
\arguments{
\item{subtype}{Floor Area grouped by: "Stock_Type", "Structure", "Function", NULL (i.e. all).}
}
\description{
Calculates relative floor area of SF, MF and NR buildings by structure type.
}
\author{
Bennet Weiss
}
