% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeBuildingFloorAreaCalibration.R
\name{calcCeBuildingFloorAreaCalibration}
\alias{calcCeBuildingFloorAreaCalibration}
\title{Calculates a factor to translate energy-related floor area to material-related floor area.
Related to from net to gross floor area.}
\usage{
calcCeBuildingFloorAreaCalibration()
}
\description{
Calculates a factor to translate energy-related floor area to material-related floor area.
Related to from net to gross floor area.
}
\author{
Bennet Weiss
}
