% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeBuildingsMI.R
\name{calcCeBuildingsMI}
\alias{calcCeBuildingsMI}
\title{Calculate Material Intensity (t/m2) of residential and commercial buildings.
Based on RASMI}
\usage{
calcCeBuildingsMI(subtype = "concrete")
}
\arguments{
\item{subtype}{Material subtype. Possible values are "concrete", "steel", "plastics".}
}
\description{
Calculate Material Intensity (t/m2) of residential and commercial buildings.
Based on RASMI
}
\author{
Bennet Weiss
}
