% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeBuildingSplit.R
\name{calcCeBuildingSplit}
\alias{calcCeBuildingSplit}
\title{Calculates relative floor area of SF, MF and NR buildings by structure type.}
\usage{
calcCeBuildingSplit()
}
\description{
Calculates relative floor area of SF, MF and NR buildings by structure type.
}
\author{
Bennet Weiss
}
