% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeBuildingsSplit.R
\name{calcCeBuildingsSplit}
\alias{calcCeBuildingsSplit}
\title{Calculates relative floor area of buildings by subtype.
Based on splits in Global Exposure Model (GEM)}
\usage{
calcCeBuildingsSplit(subtype)
}
\arguments{
\item{subtype}{Split category. Can be "Function" or "Structure"}
}
\description{
Calculates relative floor area of buildings by subtype.
Based on splits in Global Exposure Model (GEM)
}
\author{
Bennet Weiss
}
