% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeBuiltLifespan.R
\name{calcCeBuiltLifespan}
\alias{calcCeBuiltLifespan}
\title{Calculates the lifetimes of residential, commercial and industrial buildings, as well as of civil engeneering.}
\usage{
calcCeBuiltLifespan()
}
\description{
Calculates the lifetimes of residential, commercial and industrial buildings, as well as of civil engeneering.
}
\author{
Bennet Weiss
}
