% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeCaoCarbonationShare.R
\name{calcCeCaOCarbonationShare}
\alias{calcCeCaOCarbonationShare}
\title{Calculate share of cao available for carbonation.}
\usage{
calcCeCaOCarbonationShare()
}
\description{
Calculate share of cao available for carbonation.
}
\author{
Bennet Weiss
}
