% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeCaOContent.R
\name{calcCeCaOContent}
\alias{calcCeCaOContent}
\title{Calculate share of CaO in cement or clinker.}
\usage{
calcCeCaOContent(subtype)
}
\arguments{
\item{subtype}{Material of which CaO content is of interest. May be "clinker" or "CKD"}
}
\description{
Calculate share of CaO in cement or clinker.
}
\author{
Bennet Weiss
}
