% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeCarbonationRate.R
\name{calcCeCarbonationRate}
\alias{calcCeCarbonationRate}
\title{Calculate carbonation rate of different strength concrete.}
\usage{
calcCeCarbonationRate(subtype = "base")
}
\arguments{
\item{subtype}{Type of carbonation rate. May be "base", "additives", "co2", "coating", "base_buried"}
}
\description{
Calculate carbonation rate of different strength concrete.
}
\author{
Bennet Weiss
}
