% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeCementLosses.R
\name{calcCeCementLosses}
\alias{calcCeCementLosses}
\title{Calculate Losses in cement cycle.}
\usage{
calcCeCementLosses(subtype)
}
\arguments{
\item{subtype}{Loss type: can be "cement_loss_construction" or "clinker_loss_production"}
}
\description{
Calculate Losses in cement cycle.
}
\author{
Bennet Weiss
}
