% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeCKDLandfillShare.R
\name{calcCeCKDLandfillShare}
\alias{calcCeCKDLandfillShare}
\title{Calculate share of CKD that goes to landfill.}
\usage{
calcCeCKDLandfillShare()
}
\description{
Calculate share of CKD that goes to landfill.
}
\author{
Bennet Weiss
}
