% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeDwellingSplit.R
\name{calcCeDwellingSplit}
\alias{calcCeDwellingSplit}
\title{Split of residential floor area into single- (RS) and multi-family (RM) homes.}
\usage{
calcCeDwellingSplit()
}
\description{
Two-item split (sums to 1): floor area of single-family homes (RS) relative to the
total residential floor area (RS + RM), based on the Global Exposure Model (GEM).
}
\author{
Bennet Weiss
}
