% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeEndUseSplit.R
\name{calcCeEndUseSplit}
\alias{calcCeEndUseSplit}
\title{Calculates how cement consumption is split into end uses:
Res (residential), Com (commercial), Ind (industrial) and Civ (civil engineering).}
\usage{
calcCeEndUseSplit()
}
\description{
Calculates how cement consumption is split into end uses:
Res (residential), Com (commercial), Ind (industrial) and Civ (civil engineering).
}
\author{
Bennet Weiss
}
