% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeFloorspaceCorrectionFactor.R
\name{calcCeFloorspaceCorrectionFactor}
\alias{calcCeFloorspaceCorrectionFactor}
\title{Calculates a calibration factor to translate energy-related floorspace from EDGE B to material-related floorspace.}
\usage{
calcCeFloorspaceCorrectionFactor()
}
\description{
Calculates a calibration factor to translate energy-related floorspace from EDGE B to material-related floorspace.
}
\author{
Bennet Weiss
}
