% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeFloorspaceGEM.R
\name{calcCeFloorspaceGEM}
\alias{calcCeFloorspaceGEM}
\title{Calculate total floorspace of buildings from GEM, optionally grouped by subtype.}
\usage{
calcCeFloorspaceGEM(subtype = NULL, remove_ind = TRUE)
}
\arguments{
\item{subtype}{str or list(str) or None, Categories by which floorspace is resolved. Other dimensions are aggregated.
Possible entries: "Stock_Type", "Structure", "Function", NULL (i.e. all). Defaults to NULL.}

\item{remove_ind}{Logical, if TRUE (default), industrial buildings are removed from the dataset.}
}
\description{
Calculate total floorspace of buildings from GEM, optionally grouped by subtype.
}
\author{
Bennet Weiss
}
