% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeFloorspaceGHSOBAT.R
\name{calcCeFloorspaceGHSOBAT}
\alias{calcCeFloorspaceGHSOBAT}
\title{Calculate Floor Area from CeGHS-OBAT Data.
The data only provides total footprint of buildings and their average height.
Floor height data is approximated from EUBUCCO to convert to floor area.}
\usage{
calcCeFloorspaceGHSOBAT(floor_height = NULL)
}
\arguments{
\item{floor_height}{Scalar providing the manual floor height for calculation.
Defaults to NULL, in which case the average floor height is inferred from EUBUCCO data.}
}
\description{
Calculate Floor Area from CeGHS-OBAT Data.
The data only provides total footprint of buildings and their average height.
Floor height data is approximated from EUBUCCO to convert to floor area.
}
\author{
Bennet Weiss
}
