% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeHibernatingStockShare.R
\name{calcCeHibernatingStockShare}
\alias{calcCeHibernatingStockShare}
\title{Calculate share of hibernating (unused but not demolished) building stock.}
\usage{
calcCeHibernatingStockShare()
}
\description{
Calculate share of hibernating (unused but not demolished) building stock.
}
\author{
Bennet Weiss
}
