% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeIndustrializedRegions.R
\name{calcCeIndustrializedRegions}
\alias{calcCeIndustrializedRegions}
\title{Define Regions that have industrialized already.
Based on expert judgement for H12 regions.
When calling this function via calcOutput() for other than H12 region set,
use round=0 as an argument if a binary output is desired.}
\usage{
calcCeIndustrializedRegions()
}
\description{
Define Regions that have industrialized already.
Based on expert judgement for H12 regions.
When calling this function via calcOutput() for other than H12 region set,
use round=0 as an argument if a binary output is desired.
}
\author{
Bennet Weiss
}
