% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeLifetimeRelStd.R
\name{calcCeLifetimeRelStd}
\alias{calcCeLifetimeRelStd}
\title{Expert guess on the relative standard deviation of lifetime
of buildings and infrastructure}
\usage{
calcCeLifetimeRelStd()
}
\description{
Expert guess on the relative standard deviation of lifetime
of buildings and infrastructure
}
\author{
Bennet Weiss
}
