% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeMaterialApplicationSplit.R
\name{calcCeMaterialApplicationSplit}
\alias{calcCeMaterialApplicationSplit}
\title{Calculate the split of product material mass across product applications by region.}
\usage{
calcCeMaterialApplicationSplit()
}
\description{
Each material (concrete/mortar) is mapped to it's respective product applications.
}
\author{
Bennet Weiss
}
