% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeMaterialTrade.R
\name{calcCeMaterialTrade}
\alias{calcCeMaterialTrade}
\title{Calculates global clinker or cement trade (net import)
based on Chatham House, UN Comtrade, and USGS (in the case of cement).}
\usage{
calcCeMaterialTrade(subtype)
}
\arguments{
\item{subtype}{Material subtype. Can be "cement" or "clinker".}
}
\description{
Calculates global clinker or cement trade (net import)
based on Chatham House, UN Comtrade, and USGS (in the case of cement).
}
\author{
Bennet Weiss
}
