% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeProductApplicationSplit.R
\name{calcCeProductApplicationSplit}
\alias{calcCeProductApplicationSplit}
\title{Calculate strength class distribution of concrete.}
\usage{
calcCeProductApplicationSplit()
}
\description{
Calculate strength class distribution of concrete.
}
\author{
Bennet Weiss
}
