% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeProductCementContent.R
\name{calcCeProductCementContent}
\alias{calcCeProductCementContent}
\title{Calculate strength class distribution of concrete.}
\usage{
calcCeProductCementContent()
}
\description{
Calculate strength class distribution of concrete.
}
\author{
Bennet Weiss
}
