% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeProductDensity.R
\name{calcCeProductDensity}
\alias{calcCeProductDensity}
\title{Calculate the density of cement products in kg/m3.}
\usage{
calcCeProductDensity()
}
\description{
Calculate the density of cement products in kg/m3.
}
\author{
Bennet Weiss
}
