% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeProductMaterialApplicationTransform.R
\name{calcCeProductMaterialApplicationTransform}
\alias{calcCeProductMaterialApplicationTransform}
\title{Calculate which product materials correspond to which Product application}
\usage{
calcCeProductMaterialApplicationTransform()
}
\description{
Calculate which product materials correspond to which Product application
}
\author{
Bennet Weiss
}
