% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeProductMaterialSplit.R
\name{calcCeProductMaterialSplit}
\alias{calcCeProductMaterialSplit}
\title{Calculate end use product share of cement.}
\usage{
calcCeProductMaterialSplit()
}
\description{
Calculate end use product share of cement.
}
\author{
Bennet Weiss
}
