% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeProductThickness.R
\name{calcCeProductThickness}
\alias{calcCeProductThickness}
\title{Calculate strength class distribution of concrete.}
\usage{
calcCeProductThickness()
}
\description{
Calculate strength class distribution of concrete.
}
\author{
Bennet Weiss
}
