% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeStockSaturationLevel.R
\name{calcCeStockSaturationLevel}
\alias{calcCeStockSaturationLevel}
\title{Sets the regionally differentiated cement stock saturation level.
Based on expert guess informed by historic saturation levels.}
\usage{
calcCeStockSaturationLevel()
}
\description{
Sets the regionally differentiated cement stock saturation level.
Based on expert guess informed by historic saturation levels.
}
\author{
Bennet Weiss
}
