% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeStockTypeSplit.R
\name{calcCeStockTypeSplit}
\alias{calcCeStockTypeSplit}
\title{Calculates how cement consumption is split into stock types.}
\usage{
calcCeStockTypeSplit()
}
\description{
Calculates how cement consumption is split into stock types.
}
\author{
Bennet Weiss
}
