% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeStructureSplit.R
\name{calcCeStructureSplit}
\alias{calcCeStructureSplit}
\title{Calculates relative floor area of buildings by structure within each end use.
Based on splits in Global Exposure Model (GEM).}
\usage{
calcCeStructureSplit()
}
\description{
Calculates relative floor area of buildings by structure within each end use.
Based on splits in Global Exposure Model (GEM).
}
\author{
Bennet Weiss
}
