% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeTrade.R
\name{calcCeTrade}
\alias{calcCeTrade}
\title{Calculate cement/clinker trade data (imports and exports).
Based on bilateral BACI data (1995-2023).
Intra-regional trade is filtered out during aggregation.
Extended to 'target_years' by backcasting using
cement production, GDP, shipping costs and USGS trade data (for US only) as proxy.
Clinker trade is linearly faded out between 1970 and 1950.}
\usage{
calcCeTrade(subtype, category, HS = "92", target_years = NULL)
}
\arguments{
\item{subtype}{Character string specifying the scope
- Imports
- Exports}

\item{category}{Character string specifying the stage of trade
- cement
- clinker}

\item{HS}{Character string specifying the year of the HS (Harmonized System) revision of the data
- 92
- 02
- 17
- 22}

\item{target_years}{integer vector of target years for the output data.
If NULL, all years from reference (cement production) are included.
Note: the 'years' argument in calcOutput does not work properly for this function,
because backcasting of trade happens in the custom aggregation function (once the
importer + exporter dimension have been aggregated to a single region dimension to
match the reference). Since madrat checks whether the years of the returned object
cover the years parameter before aggregating, it throws a warning when years that
are obtained through backcasting are missing. The target_years parameter is a
workaround to fix these warnings - now the warning is thrown if a period is chosen
by setting target_years that is not covered by the backcasting data (i.e. production data).}
}
\value{
magpie object of the aggregated trade data
}
\description{
Calculate cement/clinker trade data (imports and exports).
Based on bilateral BACI data (1995-2023).
Intra-regional trade is filtered out during aggregation.
Extended to 'target_years' by backcasting using
cement production, GDP, shipping costs and USGS trade data (for US only) as proxy.
Clinker trade is linearly faded out between 1970 and 1950.
}
\examples{
\dontrun{
a <- calcOutput(
  type = "CeTrade", subtype = "Imports",
  category = "cement", HS = "92"
)
}
}
\seealso{
[calcOutput()]
}
\author{
Bennet Weiss, Leonie Schweiger
}
