% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeWasteSizeBound.R
\name{calcCeWasteSizeBound}
\alias{calcCeWasteSizeBound}
\title{Calculate the size bounds of waste particles for different waste types.
Based on categorization from Cao2024}
\usage{
calcCeWasteSizeBound(subtype)
}
\arguments{
\item{subtype}{Type of Bound.
Can be upper (subtype = "max") oder lower (subtype ="min") bound.}
}
\description{
Calculate the size bounds of waste particles for different waste types.
Based on categorization from Cao2024
}
\author{
Bennet Weiss
}
