% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeWasteSizeSplit.R
\name{calcCeWasteSizeSplit}
\alias{calcCeWasteSizeSplit}
\title{Calculate what size distribution concrete/mortar waste has.
Size categories defined by particle radius.}
\usage{
calcCeWasteSizeSplit()
}
\description{
Calculate what size distribution concrete/mortar waste has.
Size categories defined by particle radius.
}
\author{
Bennet Weiss
}
