% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeWasteSplit.R
\name{calcCeWasteSplit}
\alias{calcCeWasteSplit}
\title{Calculate where concrete/mortar waste goes.}
\usage{
calcCeWasteSplit()
}
\description{
Calculate where concrete/mortar waste goes.
}
\author{
Bennet Weiss
}
