% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCoGDP1900To2150.R
\name{calcCoGDP1900To2150}
\alias{calcCoGDP1900To2150}
\title{Get GDP from 1900-2150}
\usage{
calcCoGDP1900To2150(scenario = "SSP2", perCapita = FALSE)
}
\arguments{
\item{scenario}{Scenario to use for future GDP data (default: SSP2).}

\item{perCapita}{If TRUE, GDP is returned as per capita (default: FALSE).}
}
\value{
List with Magpie object of GDP (given in 2005 USD) and metadata in calcOutput format.
}
\description{
Calc GDP (PPP) from 1900-2150 yearly for the SIMSON format on a country
level. Can be aggregated to regions via calcOutput aggregate parameter.
Uses \link[=readOECD_GDP]{OECD_GDP} for historical GDP data as well as
\link[mrdrivers:calcGDP]{calcGDP} from mrdrivers for current and future GDP data
according to a specific scenario (see \code{vignette("scenarios")}
for more information). Population data from
\link[=calcCoPopulation1900To2150]{calcCoPopulation1900To2150} is used to
convert GDP per capita to total GDP.
GDP is given in 2005 USD (PPP). It's extrapolated to the past with historic
GDP datasets that use a different base year, which however does not matter
as only the relative values are used
(see \link{toolInterpolate2D}).
}
\author{
Merlin Jo Hosak
}
