% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCoPopulation.R
\name{calcCoPopulation}
\alias{calcCoPopulation}
\title{Get population from 1800-2150}
\usage{
calcCoPopulation(scenarios = "SSP2", collapse = TRUE, smooth = FALSE, dof = 8)
}
\arguments{
\item{scenarios}{Character vector or string specifying the scenarios to use for the future.}

\item{collapse}{Logical. If TRUE, redundant dimensions (e.g. scenario if only one requested) are removed.}

\item{smooth}{Logical. If TRUE, data is smoothed using spline interpolation
(see \link{toolHistoricallyConsistentSmoothing}); smoothed historical values
are identical across scenarios.}

\item{dof}{Integer. Degrees of freedom for spline interpolation.
Higher values lead to a closer fit to the original data, while lower values result in smoother curves.}
}
\value{
List with Magpie object of population and metadata in calcOutput
format.
}
\description{
Calc population from 1800-2150 yearly for the REMIND-MFA format on a country
level. Can be aggregated to regions via calcOutput aggregate parameter.
Uses \link{readGapminder} and \link[mrdrivers]{readUN_PopDiv}
datasets for historical population data as well as
\link[mrdrivers]{calcPopulation} from
mrdrivers for current and future population data for different
scenarios (see \code{vignette("scenarios")} for more information).
}
\author{
Merlin Jo Hosak, Bennet Weiss
}
