% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCullenFabricationYield.R
\name{calcCullenFabricationYield}
\alias{calcCullenFabricationYield}
\title{TODOMERLIN: document}
\usage{
calcCullenFabricationYield()
}
\description{
TODOMERLIN: document
}
\author{
Merlin Jo Hosak
}
