% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCullenStaticParameters.R
\name{calcCullenStaticParameters}
\alias{calcCullenStaticParameters}
\title{TODOMERLIN: document}
\usage{
calcCullenStaticParameters()
}
\description{
TODOMERLIN: document
}
\author{
Merlin Jo Hosak
}
