% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlBACI.R
\name{calcPlBACI}
\alias{calcPlBACI}
\title{Get BACI Trade data for specific product groups}
\usage{
calcPlBACI(subtype, category, HS)
}
\arguments{
\item{subtype}{Character string specifying the scope
- "plastics_UNCTAD": plastics trade data for HS codes from UNCTAD classification into
  primary, intermediate, manufactured, final and waste plastics
- "plastics_UNEP": plastics trade data for HS codes from UNEP NGP
  (estimated plastics mass based on estimated plastics percentages in goods)}

\item{category}{Character string specifying the stage of trade, valid parameters depend on the subtype
for plastics_UNCTAD:
- "Final"        - Final plastics
- "Primary"      - Primary plastics
- "Intermediate" - Intermediate forms of plastic
- "Manufactured" - Intermediate manufactured plastic goods
- "Application"  - Plastic goods
- "Waste"        - Plastic waste
for plastics_UNEP:
- "Primary"
- "Application"
- "Waste"}

\item{HS}{Character string specifying the year of the HS (Harmonized System) revision of the data
- 92
- 02
- 17
- 22}
}
\value{
magpie object of the BACI trade data
}
\description{
Get BACI Trade data for specific product groups
}
\examples{
\dontrun{
a <- calcOutput(
  type = "PlBACI", subtype = "plastics_UNCTAD",
  category = "Plastics in primary forms", HS = "92"
)
}
}
\seealso{
[calcOutput()]
}
\author{
Leonie Schweiger
}
