% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlBioRate.R
\name{calcPlBioRate}
\alias{calcPlBioRate}
\title{Calculate Country-Level Bio-based Plastic Share Trajectories}
\usage{
calcPlBioRate()
}
\description{
Generate time series of bio-based plastic share trajectories by sector
and aggregate from regions to countries for 1990-2100.
}
\author{
Qianzhi Zhang
}
