% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlCarbonContent.R
\name{calcPlCarbonContent}
\alias{calcPlCarbonContent}
\title{Returns the carbon contents of different plastic types/polymers or HVCs}
\usage{
calcPlCarbonContent(subtype = "plastics")
}
\arguments{
\item{subtype}{either plastics or HVC}
}
\description{
Returns the carbon contents of different plastic types/polymers or HVCs
}
\author{
Leonie Schweiger
}
