% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlChemReRate.R
\name{calcPlChemReRate}
\alias{calcPlChemReRate}
\title{Calculate Country-Level Chemical Recycling Rate Trajectories}
\usage{
calcPlChemReRate(target = NULL)
}
\description{
Generate time series of chemical recycling share trajectories by sector and region,
then aggregate to countries for 1990–2100.
}
\author{
Qianzhi Zhang
}
