% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlConsumptionByGood.R
\name{calcPlConsumptionByGood}
\alias{calcPlConsumptionByGood}
\title{Calculate Country-Level Plastic Use by Sector}
\usage{
calcPlConsumptionByGood()
}
\description{
Backcast total plastic use from 1990-2019 back to 1950.
Combine sectoral use shares and total use to compute plastic use
by sector at country level for 1950-2019.
}
\author{
Qianzhi Zhang
}
