% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlDACRate.R
\name{calcPlDACRate}
\alias{calcPlDACRate}
\title{Calculate Country-Level DAC-based Plastic Share Trajectories}
\usage{
calcPlDACRate()
}
\description{
Generate time series of Direct-Air-Capture-based plastic share trajectories by sector
and aggregate from regions to countries for 1990–2100.
}
\author{
Qianzhi Zhang
}
