% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlEoL_shares.R
\name{calcPlEoL_shares}
\alias{calcPlEoL_shares}
\title{Calculate Country-Level End-of-Life Plastic Fate Shares}
\usage{
calcPlEoL_shares(subtype)
}
\arguments{
\item{subtype}{Character string specifying the EoL treatment
Valid formats include:
- Recycled (share of total collected)
- Landfilled (share of total collected)
- Incinerated (share of total collected)
- Collected (share of total plastic waste)
- All (all shares combined in one magpie object)}
}
\description{
Compute end-of-life fate ratios of plastics by country,
based on OECD regional waste EOL data (1990–2019) and
country/region-specific datasets (EPA, Plastics Europe and Chinese reports)
}
\author{
Leonie Schweiger, Qianzhi Zhang
}
