% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlGaoCabrera.R
\name{calcPlGaoCabrera}
\alias{calcPlGaoCabrera}
\title{Gao & Cabrera-Serrenho 2025 polymer consumption in IAMC variables}
\usage{
calcPlGaoCabrera()
}
\value{
Total polymer consumption (Mt/yr)
}
\description{
Return total apparent polymer consumption from Gao & Cabrera-Serrenho (2025),
"Global Plastic MFA", at ISO3 country level. The 14 polymer groups (plastics,
fibres and rubber) are summed into a single total and mapped to the IAMC
variable \code{Material Demand|Chemicals|Plastics} in Mt/yr. Note that the
years 2020-2021 rely on incomplete trade data.
}
\seealso{
\code{\link{readGaoCabrera2025}}, \code{\link{calcPlPottinger}}
}
\author{
Leonie Schweiger
}
