% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlHVCinput.R
\name{calcPlHVCinput}
\alias{calcPlHVCinput}
\title{Calculates the parameters that determine the HVC input to virgin plastics production
based on data by Levi and Cullen (2018) https://doi.org/10.1021/acs.est.7b04573}
\usage{
calcPlHVCinput(subtype)
}
\arguments{
\item{subtype}{Character string specifying the parameter to read
- "polymerization_yield"
- "HVC_input_ratio"
- "C4_input_ratio"}
}
\description{
Calculates the parameters that determine the HVC input to virgin plastics production
based on data by Levi and Cullen (2018) https://doi.org/10.1021/acs.est.7b04573
}
\author{
Leonie Schweiger
}
