% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlMechLoss.R
\name{calcPlMechLoss}
\alias{calcPlMechLoss}
\title{Set Country-Level Mechanical Recycling Loss}
\usage{
calcPlMechLoss()
}
\description{
Set Country-Level Mechanical Recycling Loss
}
\author{
Qianzhi Zhang
}
