% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlMechReRate.R
\name{calcPlMechReRate}
\alias{calcPlMechReRate}
\title{Calculate Country-Level Mechanical Recycling Rate Trajectories}
\usage{
calcPlMechReRate()
}
\description{
Generate time series of mechanical recycling rates by sector and region
using OECD end-of-life (EoL) data and other sources, then aggregate to countries for 1990–2100.
}
\author{
Qianzhi Zhang
}
