% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlOECD_EoL.R
\name{calcPlOECD_EoL}
\alias{calcPlOECD_EoL}
\title{Calculate Country-Level End-of-Life Plastic Fate Shares}
\usage{
calcPlOECD_EoL()
}
\description{
Compute end-of-life fate ratios of plastics by country,
based on OECD regional waste EOL data (1990–2019).
}
\author{
Qianzhi Zhang
}
