% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlPlasticsEurope.R
\name{calcPlPlasticsEurope}
\alias{calcPlPlasticsEurope}
\title{This function returns plastics production data from Plastics Europe}
\usage{
calcPlPlasticsEurope()
}
\description{
This function returns plastics production data from Plastics Europe
}
\author{
Leonie Schweiger
}
