% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlSectorSplit.R
\name{calcPlSectorSplit}
\alias{calcPlSectorSplit}
\title{Returns the plastics sector splits from Ryberg et al. 2019,
mapping Ryberg sectors to remind-mfa sectors}
\usage{
calcPlSectorSplit()
}
\description{
Returns the plastics sector splits from Ryberg et al. 2019,
mapping Ryberg sectors to remind-mfa sectors
}
\author{
Leonie Schweiger
}
