% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlUNCTAD.R
\name{calcPlUNCTAD}
\alias{calcPlUNCTAD}
\title{Calculate UNCTAD Plastic Trade Flows}
\usage{
calcPlUNCTAD(subtype)
}
\arguments{
\item{subtype}{Character; scenario to read. Options:
\itemize{
  \item Final        - Imports/exports of final plastics by country
  \item Waste        - Plastic waste imports/exports by country
  \item Primary      - Imports/exports of primary plastics by country
  \item Intermediate - Imports/exports of intermediate forms of plastic by country
  \item Manufactured - Imports/exports of intermediate manufactured plastic goods by country
}}
}
\description{
Read and aggregate UNCTAD plastic trade data (imports/exports) for multiple plastic categories
at country level.
}
\author{
Qianzhi Zhang
}
