% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlZeroRates.R
\name{calcPlZeroRates}
\alias{calcPlZeroRates}
\title{Dummy function to create magpie object with rates that are historically 0.
E.g. for chemical recycling, bio-based production and DAC based production}
\usage{
calcPlZeroRates()
}
\description{
Dummy function to create magpie object with rates that are historically 0.
E.g. for chemical recycling, bio-based production and DAC based production
}
\author{
Leonie Schweiger
}
