% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStBACI.R
\name{calcStBACI}
\alias{calcStBACI}
\title{Split World Steel trade into bilateral trade by using proxy bilateral BACI trade data
and aggregate via a custom aggregation function to filter out intra-regional trade
(since we don't have a list with all HS codes and corresponding steel shares,
BACI trade data cannot be used directly, since it does not cover the full scope)}
\usage{
calcStBACI(subtype, category, HS = "92")
}
\arguments{
\item{subtype}{Character string specifying the scope
- imports
- exports}

\item{category}{Character string specifying the stage of trade
- direct
- indirect
- scrap}

\item{HS}{Character string specifying the year of the HS (Harmonized System) revision of the data
- 92
- 02
- 17
- 22}
}
\value{
magpie object of the aggregated trade data
}
\description{
Split World Steel trade into bilateral trade by using proxy bilateral BACI trade data
and aggregate via a custom aggregation function to filter out intra-regional trade
(since we don't have a list with all HS codes and corresponding steel shares,
BACI trade data cannot be used directly, since it does not cover the full scope)
}
\examples{
\dontrun{
a <- calcOutput(type = "StBACI", subtype = "imports",
category = "direct", HS = "02")
}
}
\seealso{
[calcOutput()]
}
\author{
Leonie Schweiger
}
