% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStCullenFabricationYield.R
\name{calcStCullenFabricationYield}
\alias{calcStCullenFabricationYield}
\title{Load Cullen 2012 fabrication yield}
\usage{
calcStCullenFabricationYield()
}
\description{
Function to load fabrication yield data from Cullen et al. (2012).
Using the function \link{readCullen2012}.
}
\author{
Merlin Jo Hosak
}
