% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStCullenStaticParameters.R
\name{calcStCullenStaticParameters}
\alias{calcStCullenStaticParameters}
\title{TODOMERLIN: document}
\usage{
calcStCullenStaticParameters()
}
\description{
TODOMERLIN: document
}
\author{
Merlin Jo Hosak
}
