% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStDRIData.R
\name{calcStDRIData}
\alias{calcStDRIData}
\title{Calc DRI data}
\usage{
calcStDRIData(subtype)
}
\arguments{
\item{subtype}{Options: 'production', 'imports', 'exports'}
}
\description{
Calc direct reduced iron data mostly based on World Steel Association
data (see \link{readWorldSteelDatabase} for details).
The DRI data is backcasted based on steel production data to fill in missing
years.
}
\author{
Merlin Jo Hosak
}
