% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSteelLifetimes.R
\name{calcSteelLifetimes}
\alias{calcSteelLifetimes}
\title{TODOMERLIN: document}
\usage{
calcSteelLifetimes(subtype)
}
\arguments{
\item{subtype}{TODOMERLIN: document}
}
\description{
TODOMERLIN: document
}
\author{
Merlin Jo Hosak
}
