% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSteelMaxScrapShare.R
\name{calcSteelMaxScrapShare}
\alias{calcSteelMaxScrapShare}
\title{Calc max steel scrap share in production}
\usage{
calcSteelMaxScrapShare(subtype = "BIR")
}
\arguments{
\item{subtype}{Data source used, current default is BIR (Bureau of
International Recycling).}
}
\description{
Estimate the maximum share steel scrap can have in steel production
based on data by the Bureau of International Recycling (BIR),
see \link{readBIR} and the folder 'BIR' in sources.
}
\author{
Merlin Jo Hosak
}
