% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSteelProductionByProcess.R
\name{calcSteelProductionByProcess}
\alias{calcSteelProductionByProcess}
\title{Calc Steel Production by Process}
\usage{
calcSteelProductionByProcess(
  assumedPastPercentages = list(y1900 = c(0.2, 0, 0.8))
)
}
\arguments{
\item{assumedPastPercentages}{A list of assumed percentages for BOF, EAF,
and Other processes for specific missing years. By default, only year 1900 is
set to c(0.2, 0, 0.8) (20% BOF, 0% EAF, 80% Other). Currently, data before
1982 is only based on these assumptions}
}
\value{
Steel production by process across all regions from 1900-2022 as magpie within
list of metadata (in calcOutput format).
}
\description{
Calc steel production by process from WorldSteel datasets. Can be aggregated to regions
via calcOutput aggregate parameter. Uses
\link{readWorldSteelDigitised} and
\link{readWorldSteelDatabase} datasets, the former for
historic, the latter for current data. Out of this, the shares of BOF,
EAF and Other processes are calculated and interpolated/extrapolated to cover
the full time period from 1900-2022. \link{calcStProduction}
is used to then multiply the percentages as this data is likely more reliable
whilst the data split from the process data is still informative.
}
\author{
Merlin Jo Hosak
}
