% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSteelRecoveryRate.R
\name{calcSteelRecoveryRate}
\alias{calcSteelRecoveryRate}
\title{TODOMERLIN: document}
\usage{
calcSteelRecoveryRate(subtype)
}
\arguments{
\item{subtype}{TODOMERLIN: document}
}
\description{
TODOMERLIN: document
}
\author{
Merlin Jo Hosak
}
