% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSteelSectorSplits.R
\name{calcSteelSectorSplits}
\alias{calcSteelSectorSplits}
\title{TODOMERLIN: document}
\usage{
calcSteelSectorSplits(subtype)
}
\arguments{
\item{subtype}{TODOMERLIN: document}
}
\description{
TODOMERLIN: document
}
\author{
Merlin Jo Hosak
}
