% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSteelStaticParameters.R
\name{calcSteelStaticParameters}
\alias{calcSteelStaticParameters}
\title{TODOMERLIN: document}
\usage{
calcSteelStaticParameters()
}
\description{
TODOMERLIN: document
}
\author{
Merlin Jo Hosak
}
