% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStManufacturedProductsTradeShares.R
\name{calcStManufacturedProductsTradeShares}
\alias{calcStManufacturedProductsTradeShares}
\title{Calculate indirect trade shares from World Steel Digitised data on indirect
trade in 2013.}
\usage{
calcStManufacturedProductsTradeShares()
}
\description{
Calculate indirect trade shares from World Steel Digitised data on indirect
trade in 2013.
}
\author{
Falk Benke
}
