% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStLifetimes.R
\name{calcStLifetimes}
\alias{calcStLifetimes}
\title{Calc Steel Lifetimes}
\usage{
calcStLifetimes(subtype, unit)
}
\arguments{
\item{subtype}{Subtype of steel lifetimes data to load. Currently, only
'Cooper2014' is available. Other options might include Pauliuk or Wittig
(see respective folders).}

\item{unit}{'mean' to return mean lifetimes, 'std' to return standard deviation}
}
\description{
Function to load steel lifetimes data in years.
}
\author{
Merlin Jo Hosak
}
