% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStPigIronProduction.R
\name{calcStPigIronProduction}
\alias{calcStPigIronProduction}
\title{Calc pig iron production}
\usage{
calcStPigIronProduction()
}
\description{
Calculate pig iron production. Wrapper function. See
\link{calcStPigIronPreliminaryData} for specifics.
}
\author{
Merlin Jo Hosak
}
