% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStPigIronTrade.R
\name{calcStPigIronTrade}
\alias{calcStPigIronTrade}
\title{Calc pig iron trade}
\usage{
calcStPigIronTrade(subtype)
}
\arguments{
\item{subtype}{'imports' or 'exports' to calculate pig iron imports or
exports respectively.}
}
\description{
Calculate pig iron trade Wrapper function. See
\link{calcStPigIronPreliminaryData} for specifics.
}
\author{
Merlin Jo Hosak
}
