% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStScrapConsumption.R
\name{calcStScrapConsumption}
\alias{calcStScrapConsumption}
\title{Calc steel scrap consumption}
\usage{
calcStScrapConsumption(subtype, aggregate = NULL, regionmapping = NULL)
}
\arguments{
\item{subtype}{Subtype of steel scrap consumption data to calculate.
Options: 'assumptions' or 'noAssumptions'. Must be set deliberately.}

\item{aggregate}{Aggregation level. TRUE uses regionmapping,
F for country level, 'GLO' for global level. See \link[madrat:calcOutput]{calcOutput}
documentation. ++REGGLO/regglo/reg+glo etc. deprecated.}

\item{regionmapping}{Regionmapping to use for regional aggregation. Should default
to H12 REMIND regions.}
}
\description{
Function to calculate steel scrap consumption based on World Steel
Association (see \link{readWorldSteelDigitised})
data and Bureau of International Recycling (BIR) (see
\link{readBIR}) data. Different subtypes allow to either use
assumptions to fill data gaps or not.

With assumptions, basic backcasting by production data and forecasting
via BIR data is used. For no assumptions, only linear interpolation is used.
Data is manipulated so that when aggregating regions with missing data
in less relevant subunits, these missing data are set to 0 to avoid NAs in
the regional aggregation (as well as global aggregation).

The countries were NA should be zeroed are defined in
scrap_consumption_countries_2_zero.csv .
}
\author{
Merlin Jo Hosak
}
