% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStScrapConsumptionWS.R
\name{calcStScrapConsumptionWS}
\alias{calcStScrapConsumptionWS}
\title{Calculate scrap consumption based on source WorldSteelDigitised}
\usage{
calcStScrapConsumptionWS()
}
\description{
Calculate scrap consumption based on source WorldSteelDigitised
}
\author{
Falk Benke
}
