% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStTrade.R
\name{calcStTrade}
\alias{calcStTrade}
\title{Split World Steel trade into bilateral trade by using proxy bilateral BACI trade data
and aggregate via a custom aggregation function to filter out intra-regional trade
(since we don't have a list with all HS codes and corresponding steel shares,
BACI trade data cannot be used directly, since it does not cover the full scope)}
\usage{
calcStTrade(
  subtype,
  category,
  HS = "92",
  include_intra_regional = FALSE,
  target_years = NULL
)
}
\arguments{
\item{subtype}{Character string specifying the scope
- imports
- exports}

\item{category}{Character string specifying the stage of trade
- direct
- indirect
- scrap}

\item{HS}{Character string specifying the year of the HS (Harmonized System) revision of the data
- 92
- 02
- 17
- 22}

\item{include_intra_regional}{bool if intra-regional trade should be included}

\item{target_years}{integer vector of target years for the output data.
If NULL, all years from reference are included.
Note: the 'years' argument in calcOutput does not work properly for this function,
because backcasting of trade happens in the custom aggregation function (once the
importer + exporter dimension have been aggregated to a single region dimension to
match the reference). Since madrat checks whether the years of the returned object
cover the years parameter before aggregating, it throws a warning when years that
are obtained through backcasting are missing. The target_years parameter is a
workaround to fix these warnings - now the warning is thrown if a period is chosen
by setting target_years that is not covered by the backcasting data (i.e. production data).}
}
\value{
magpie object of the aggregated trade data
}
\description{
Split World Steel trade into bilateral trade by using proxy bilateral BACI trade data
and aggregate via a custom aggregation function to filter out intra-regional trade
(since we don't have a list with all HS codes and corresponding steel shares,
BACI trade data cannot be used directly, since it does not cover the full scope)
}
\examples{
\dontrun{
a <- calcOutput(
  type = "StTrade", subtype = "imports",
  category = "direct", HS = "02"
)
}
}
\seealso{
[calcOutput()]
}
\author{
Leonie Schweiger
}
