% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStTradeWorldsteel.R
\name{calcStTradeWorldsteel}
\alias{calcStTradeWorldsteel}
\title{Get steel trade data}
\usage{
calcStTradeWorldsteel(subtype = "imports")
}
\arguments{
\item{subtype}{Type of trade data to retrieve. Options: "imports", "exports",
"scrapImports", "scrapExports", "manufacturedProductsImports", "manufacturedProductsExports"}
}
\value{
Steel trade across all regions from 1900-2022 as magpie within
list of metadata.
}
\description{
Calc steel trade from WorldSteel datasets. Can be aggregated to regions
via calcOutput aggregate parameter. Uses
\link{readWorldSteelDigitised} and
\link{readWorldSteelDatabase} datasets, the former for
historic, the latter for current data. Further, uses
\link{calcStProduction} to backcast historic trade data.
}
\author{
Merlin Jo Hosak
}
