% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStWorldSteelStaticParameters.R
\name{calcStWorldSteelStaticParameters}
\alias{calcStWorldSteelStaticParameters}
\title{Calc World Steel Static Parameters}
\usage{
calcStWorldSteelStaticParameters(subtype)
}
\arguments{
\item{subtype}{Parameter to load. Currently only 'scrapInBOFRate' available.}
}
\description{
Load static (singular) parameters based on World Steel Association data.
See \link{readWorldSteelParameters} and the
WorldSteelParameters folder for more information.
}
\author{
Merlin Jo Hosak
}
