% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEUBUCCO.R
\name{convertEUBUCCO}
\alias{convertEUBUCCO}
\title{Convert data from EUBUCCO}
\usage{
convertEUBUCCO(x)
}
\arguments{
\item{x}{Magpie object}
}
\description{
Convert data from EUBUCCO
}
\author{
Bennet Weiss
}
