% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGapminder.R
\name{convertGapminder}
\alias{convertGapminder}
\title{Convert Gapminder data to Magpie format}
\usage{
convertGapminder(x, subtype)
}
\arguments{
\item{x}{MagPIE object}

\item{subtype}{Regional scope of the data. Either "countries" or "global".}
}
\description{
Convert Gapminder data to Magpie format
}
\author{
Merlin Jo Hosak, Bennet Weiss
}
