% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGEM.R
\name{convertGEM}
\alias{convertGEM}
\title{Convert buildings split data from GEM.}
\usage{
convertGEM(x)
}
\arguments{
\item{x}{Magpie object}
}
\description{
Convert buildings split data from GEM.
}
\author{
Bennet Weiss
}
