% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGHSOBAT.R
\name{convertGHSOBAT}
\alias{convertGHSOBAT}
\title{Convert GHSOBAT}
\usage{
convertGHSOBAT(x, subtype)
}
\arguments{
\item{x}{magpie object}
}
\description{
Convert GHSOBAT
}
\author{
Bennet Weiss

subtype Data to extract.
       "surface" for total footprint building differentiated by residential and non-residential.
       "height" for average building height (not differentiated)
}
