% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGlobalExposureModel.R
\name{convertGlobalExposureModel}
\alias{convertGlobalExposureModel}
\title{Convert buildings split data from GEM.}
\usage{
convertGlobalExposureModel(x)
}
\arguments{
\item{x}{Magpie object}
}
\description{
Convert buildings split data from GEM.
}
\author{
Bennet Weiss
}
