% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEDC.R
\name{convertIEDC}
\alias{convertIEDC}
\title{Read IEDC data}
\usage{
convertIEDC(x, subtype)
}
\arguments{
\item{x}{MagPIE object}

\item{subtype}{Options: 'pigIronProduction', 'pigIronImports', 'pigIronExports'}
}
\description{
Read data from the Industrial Ecology Data Commons (IEDC). So far
only data by Pauliuk et al. (2019) on pig iron is used.
}
\author{
Falk Benke
}
