% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIRSG.R
\name{convertIRSG}
\alias{convertIRSG}
\title{Convert IRSG synthetic rubber data to a magpie object at country level}
\usage{
convertIRSG(x, subtype)
}
\arguments{
\item{x}{magpie object of IRSG regional data.}

\item{subtype}{Character string, must be "regional".}
}
\value{
magpie object of the IRSG data disaggregated to country level.
}
\description{
Disaggregates the IRSG regional synthetic rubber production to ISO country
level. The regions (Asia, Europe, Americas, Africa) are split into their
member countries (see \code{regionmappingIRSG.csv}) weighted by country-level
chemical energy consumption from \code{\link[mrindustry]{calcChemicalTotal}}.
}
\details{
Only the "regional" subtype is country-resolved. The "global_total" subtype
is global data and must be read with \code{convert = FALSE}.
}
\author{
Leonie Schweiger
}
