% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertOECDAffordableHousingDatabase.R
\name{convertOECDAffordableHousingDatabase}
\alias{convertOECDAffordableHousingDatabase}
\title{Convert OECD Affordable Housing Database.}
\usage{
convertOECDAffordableHousingDatabase(x)
}
\arguments{
\item{x}{Magpie object}
}
\description{
Convert OECD Affordable Housing Database.
}
\author{
Bennet Weiss
}
