% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertStegmann2022.R
\name{convertStegmann2022}
\alias{convertStegmann2022}
\title{Convert Stegmann 2022 plastics data to ISO country level}
\usage{
convertStegmann2022(x)
}
\arguments{
\item{x}{MagPIE object with Stegmann2022 data at the 26-region resolution.}
}
\value{
MagPIE object with the Stegmann2022 data disaggregated to ISO3 country
level (production in PJ/yr, population in million).
}
\description{
Disaggregate the 26 Stegmann (IMAGE) regions to ISO3 country level. The
regional production (PJ/yr) and population (million) are distributed across the
member countries of each region proportionally to population from
\code{calcOutput("CoPopulation", scenarios = "SSP2")}, whose yearly 1800-2150
coverage spans the Stegmann years (2005-2100).
}
\seealso{
\code{\link{readStegmann2022}}
}
\author{
Leonie Schweiger
}
