% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertUNCTAD.R
\name{convertUNCTAD}
\alias{convertUNCTAD}
\title{Convert UNCTAD}
\usage{
convertUNCTAD(x)
}
\arguments{
\item{x}{MagPIE object containing UNCTAD data}
}
\value{
MagPIE object of the UNCTAD data mapped to ISO countries
}
\description{
Convert UNCTAD data to ISO country level.
}
\examples{
\dontrun{
a <- convertUNCTAD(x)
}

}
\author{
Leonie Schweiger
}
