% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertWorldSteelDigitised.R
\name{convertWorldSteelDigitised}
\alias{convertWorldSteelDigitised}
\title{Convert data World Steel Association digitised 1978-2022 yearbooks.}
\usage{
convertWorldSteelDigitised(x, subtype = "production")
}
\arguments{
\item{x}{Magpie object}

\item{subtype}{TODOMERLIN: document}
}
\description{
Convert data World Steel Association digitised 1978-2022 yearbooks.
}
\author{
Merlin Jo Hosak
}
