% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertXi2016.R
\name{convertXi2016}
\alias{convertXi2016}
\title{Convert data from Xi2016}
\usage{
convertXi2016(x)
}
\arguments{
\item{x}{Magpie object}
}
\description{
Convert data from Xi2016
}
\author{
Bennet Weiss
}
