% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullMFA.R
\name{fullMFA}
\alias{fullMFA}
\title{Load all MFA data}
\usage{
fullMFA(
  rev = 0,
  dev = "",
  gdpPerCapita = TRUE,
  driverScenario = c("SSP1", "SSP2", "SSP3", "SSP4", "SSP5"),
  runSections = NULL,
  end_future = 2100
)
}
\arguments{
\item{rev}{Revision number for the data version}

\item{dev}{Development version string}

\item{gdpPerCapita}{bool if GDP should be returned as per capita values.}

\item{driverScenario}{Character vector specifying the scenario to use for GDP and population data.}

\item{runSections}{Character vector selecting which parts to run.
Allowed values (see validSections): c("steel", "cement", "plastic"). NULL (default) runs all.}

\item{end_future}{End year for future data (default: 2100).}
}
\description{
Function that produces the complete regional data sets required for the
MFA model.
}
\examples{
\dontrun{
retrieveData("MFA")
fullMFA()
}

}
\seealso{
\code{\link[madrat]{readSource}}, \code{\link[madrat]{getCalculations}},
\code{\link[madrat]{calcOutput}}
}
\author{
Merlin HOSAK

Bennet Weiss
}
