% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSources_mrmfa.R
\name{getSources_mrmfa}
\alias{getSources_mrmfa}
\title{Get Sources mrmfa}
\usage{
getSources_mrmfa()
}
\description{
Collects the sources of all calcFunctions that are called in fullMFA
and their respective bibtex entries in SOURCE_INFO.txt if available
Saves a table with the calcFunctions, source names and bibtex keys as csv
Saves all bibtex entries in a .bib file
}
\author{
Leonie Schweiger
}
