% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeFloorspaceCorrectionFactor.R
\name{plot_floor_area_comparison}
\alias{plot_floor_area_comparison}
\title{Plots a comparison of floor area data from EDGE-B, EUBUCCO and GEM.}
\usage{
plot_floor_area_comparison(
  edgeb_floor_area,
  eubucco_floor_area,
  gem_floor_area,
  ghsoobat_floor_area
)
}
\arguments{
\item{edgeb_floor_area}{EDGE-B floor area magpie object.}

\item{eubucco_floor_area}{EUBUCCO floor area magpie object.}

\item{gem_floor_area}{GEM floor area magpie object.}

\item{ghsoobat_floor_area}{GHS-OBAT floor area magpie object.}
}
\description{
Plots a comparison of floor area data from EDGE-B, EUBUCCO and GEM.
}
\author{
Bennet Weiss
}
