% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeFloorspaceCorrectionFactor.R
\name{plot_floorspace_data}
\alias{plot_floorspace_data}
\title{Generates different plots of floor area data.}
\usage{
plot_floorspace_data(plotting)
}
\arguments{
\item{plotting}{Type of plot to generate.
Options are "floor area comparison" and "ratio eubucco/edgeb over cement production".
Defaults to no plot.}
}
\description{
Generates different plots of floor area data.
}
\author{
Bennet Weiss
}
