% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeFloorspaceCorrectionFactor.R
\name{plot_ratio_over_x}
\alias{plot_ratio_over_x}
\title{Plots edge_b / eubucco ratio as function of given x variable.}
\usage{
plot_ratio_over_x(edgeb, eubucco, x, xlabel)
}
\arguments{
\item{edgeb}{EDGE-B floor area magpie object}

\item{eubucco}{EUBUCCO floor area magpie object}

\item{x}{magpie object to plot ratio against (e.g., Production, GDP)}

\item{xlabel}{label for x axis}
}
\description{
Plots edge_b / eubucco ratio as function of given x variable.
}
\author{
Bennet Weiss
}
