% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBACI.R
\name{readBACI}
\alias{readBACI}
\title{Read BACI Trade data}
\usage{
readBACI(subtype, subset)
}
\arguments{
\item{subtype}{Character string specifying the scope, combination of material HS codes and category
material HS codes:
- "plastics_UNCTAD": plastics trade data for HS codes from UNCTAD classification into
  primary, intermediate, manufactured, final and waste plastics
  valid parameters for category:
  - Plastics in primary forms
  - Intermediate forms of plastic
  - Intermediate manufactured plastic goods
  - Final manufactured plastics goods
  - Plastic waste
- "plastics_UNEP": plastics trade data for HS codes from UNEP NGP
  (estimated plastics mass based on estimated plastics percentages in goods)
  valid parameters for category:
  - Primary
  - Application
  - Waste
- "steel": steel trade data from custom mapping
  valid parameters for category:
  - direct
  - indirect
  - scrap
- "cement": cement trade data
  valid parameters for category:
  - cement
  - clinker}

\item{subset}{Character string specifying the HS (Harmonized System) revision of the data
- 92
- 02
- 17
- 22}
}
\value{
magpie object of the BACI trade data
}
\description{
Read BACI Trade data
}
\examples{
\dontrun{
a <- readSource(type = "BACI", subtype = "plastics_UNEP-Primary", subset = "92")
}
}
\seealso{
[readSource()]
}
\author{
Leonie Schweiger
}
