% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBIR.R
\name{readBIR}
\alias{readBIR}
\title{Read BIR (Bureau of International Recycling) data}
\usage{
readBIR(subtype)
}
\arguments{
\item{subtype}{Type of data to read. Options: "scrapConsumption"}
}
\description{
Reads BIR data on steel scrap consumption.
}
\author{
Merlin Jo Hosak
}
