% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCao2024.R
\name{readCao2024}
\alias{readCao2024}
\title{Read data received on 04.08.2025, personal communication.
Data used for Kaufmann et al. (2024), DOI: 10.1088/1748-9326/ad236b
"Society's material stocks as carbon pool: an economy-wide quantification of global carbon stocks from 1900-2015"
The data is given as probability density functions.
For further use, they are translated to means before turned to magclass object.}
\usage{
readCao2024(subtype)
}
\arguments{
\item{subtype}{Variable to be read in.
Supported subtypes are:
"product_material_split", "product_application_split", "carbonation_rate", "carbonation_rate_factor_additives",
"carbonation_rate_factor_co2", "carbonation_rate_factor_coating", "carbonation_rate_buried", "product_thickness",
"product_cement_content", "cao_carbonation_share", "cement_loss_construction", "clinker_loss_production",
"waste_split", "waste_size_split", "CKD_landfill_share", "clinker_cao_content", "CKD_cao_content"}
}
\description{
Read data received on 04.08.2025, personal communication.
Data used for Kaufmann et al. (2024), DOI: 10.1088/1748-9326/ad236b
"Society's material stocks as carbon pool: an economy-wide quantification of global carbon stocks from 1900-2015"
The data is given as probability density functions.
For further use, they are translated to means before turned to magclass object.
}
\author{
Bennet Weiss
}
