% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readChina_PlasticEoL.R
\name{readChina_PlasticEoL}
\alias{readChina_PlasticEoL}
\title{Read China_PlasticEoL Data into a magpie Object}
\usage{
readChina_PlasticEoL()
}
\value{
magpie object of the China Plastics EoL Data
}
\description{
This function reads data on plastics EoL treatment in China from different sources
}
\seealso{
[readSource()]

\dontrun{
a <- readSource(type = "China_PlasticEoL")
}
}
\author{
Leonie Schweiger
}
