% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCullen2012.R
\name{readCullen2012}
\alias{readCullen2012}
\title{TODOMERLIN: document}
\usage{
readCullen2012(subtype)
}
\arguments{
\item{subtype}{TODOMERLIN: document}
}
\description{
TODOMERLIN: document
}
\author{
Merlin Jo Hosak
}
