% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEUBUCCO.R
\name{readEUBUCCO}
\alias{readEUBUCCO}
\title{Read data from EUBUCCO
This is preliminary and unpublished data.
Data received on 30.09.2025, PIK-internal personal communication.}
\usage{
readEUBUCCO(subtype = "concrete")
}
\arguments{
\item{subtype}{Variable to be read in. Currently, only "clinker_ratio" is supported.}
}
\description{
Read data from EUBUCCO
This is preliminary and unpublished data.
Data received on 30.09.2025, PIK-internal personal communication.
}
\author{
Bennet Weiss
}
