% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGapminder.R
\name{readGapminder}
\alias{readGapminder}
\title{Read Gapminder population data.}
\usage{
readGapminder(subtype = "population")
}
\arguments{
\item{subtype}{TODOMERLIN: document}
}
\description{
Read Gapminder population data from 1800-2100 in yearly resolution.
}
\author{
Merlin Jo Hosak
}
