% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGeyer.R
\name{readGeyer}
\alias{readGeyer}
\title{Read Geyer et al. SI Global plastic production 1950-2015}
\usage{
readGeyer(subtype)
}
\arguments{
\item{subtype}{Character string specifying the dataset and scope
- "Prod_1950-2015"
- "Lifetime_mean"
- "Lifetime_std"}
}
\value{
magpie object of the Geyer et al. Plastic data
}
\description{
Read Geyer et al. SI Global plastic production 1950-2015
}
\seealso{
[readSource()]

\dontrun{
a <- readSource(type = "Geyer", subtype = "Prod_1950-2015")
}
}
\author{
Leonie Schweiger
}
