% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGHSOBAT.R
\name{readGHSOBAT}
\alias{readGHSOBAT}
\title{Read GHS-OBAT: Global Human Settlement - Open Buildings Attribute Table
Florio, Pietro; Politis, Panagiotis; Goch, Katarzyna; Uhl, Johannes H; Melchiorri, Michele; Pesaresi, Martino; Kemper, Thomas (2024):
GHS-OBAT R2024A - Global Open Building Attribute Table at footprint level, with age, function, height and compactness information (2020).
European Commission, Joint Research Centre (JRC) [Dataset] doi: 10.2905/f41a22f1-5741-4c41-86eb-6384654f6927
PID: http://data.europa.eu/89h/f41a22f1-5741-4c41-86eb-6384654f6927
Global countrystats downloaded from
https://jeodpp.jrc.ec.europa.eu/ftp/jrc-opendata/GHSL/GHS_OBAT_GLOBE_R2024A/GHS_OBAT_COUNTRYSTATS_GLOBE_R2024A/V1-0/}
\usage{
readGHSOBAT(subtype)
}
\arguments{
\item{subtype}{Data to extract.
"surface" for total footprint building differentiated by residential and non-residential.
"height" for average building height (not differentiated)}
}
\description{
Read GHS-OBAT: Global Human Settlement - Open Buildings Attribute Table
Florio, Pietro; Politis, Panagiotis; Goch, Katarzyna; Uhl, Johannes H; Melchiorri, Michele; Pesaresi, Martino; Kemper, Thomas (2024):
GHS-OBAT R2024A - Global Open Building Attribute Table at footprint level, with age, function, height and compactness information (2020).
European Commission, Joint Research Centre (JRC) [Dataset] doi: 10.2905/f41a22f1-5741-4c41-86eb-6384654f6927
PID: http://data.europa.eu/89h/f41a22f1-5741-4c41-86eb-6384654f6927
Global countrystats downloaded from
https://jeodpp.jrc.ec.europa.eu/ftp/jrc-opendata/GHSL/GHS_OBAT_GLOBE_R2024A/GHS_OBAT_COUNTRYSTATS_GLOBE_R2024A/V1-0/
}
\author{
Bennet Weiss
}
