% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEDC.R
\name{readIEDC}
\alias{readIEDC}
\title{Read IEDC data}
\usage{
readIEDC(subtype = "pigIronProduction")
}
\arguments{
\item{subtype}{Options: 'pigIronProduction', 'pigIronImports', 'pigIronExports'}
}
\description{
Read data from the Industrial Ecology Data Commons (IEDC). So far
only data by Pauliuk et al. (2019) on pig iron is used.
}
\author{
Merlin Jo Hosak
}
