% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIRSG.R
\name{readIRSG}
\alias{readIRSG}
\title{Read IRSG synthetic rubber data into a magpie object}
\usage{
readIRSG(subtype)
}
\arguments{
\item{subtype}{Character string specifying the dataset:
- "global_total": global synthetic rubber production in Mt by year
  (year, global_SR_production_Mt). Global data (region GLO).
- "regional": regional synthetic rubber production in Mt by region and
  year (region, year, regional_SR_production_Mt).
  Regions are Asia, Europe, Americas and Africa (IRSG grouping).}
}
\value{
magpie object of the IRSG synthetic rubber data
}
\description{
This function reads International Rubber Study Group (IRSG) data on synthetic
rubber (SR) production from an Excel file, based on a specified subtype, and
returns a magpie object.
}
\examples{
\dontrun{
a <- readSource(type = "IRSG", subtype = "regional", convert = FALSE)
}
}
\seealso{
[readSource()]
}
\author{
Leonie Schweiger
}
