% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOECD_GDP.R
\name{readOECD_GDP}
\alias{readOECD_GDP}
\title{Read OECD GDP per capita data from 1500-2016}
\usage{
readOECD_GDP(subtype = "gdppc")
}
\description{
Read OECD GDP per capita data from 1500-2016
}
\author{
Merlin Jo Hosak
}
