% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOECD_Plastic.R
\name{readOECD_Plastic}
\alias{readOECD_Plastic}
\title{Read OECD Plastic Data into a magpie Object}
\usage{
readOECD_Plastic(subtype)
}
\arguments{
\item{subtype}{Character string specifying the dataset and scope.
Valid formats include:
- "Use_2019_region" (plastic use in Mt, dimensions: region, application, polymer)
- "Use_1990-2019_region" (plastic use in Mt, dimensions: region, year)
- "Use_1990-2019_world" (plastic use in Mt, dimensions: year, application, polymer, type (primary/secondary))
- "WasteEOL_1990-2019_region" (plastic waste in Mt, dimensions: region, year, EOL fate, collected for recycling)
- "WasteType_2019_region" (plastic waste in Mt, dimensions: region, application, polymer)
- "Use_1980-2060_projection" (OECD Global Plastics Outlook 2022 plastics use in Mt,
  dimensions: OECD-Outlook region, year)
- "Pop_2019-2060_projection" (OECD Global Plastics Outlook 2022 population in billion
  people, dimensions: OECD-Outlook region, year)}
}
\value{
magpie object of the OECD Plastic data
}
\description{
This function reads OECD plastic use or waste data from Excel files,
based on a specified subtype, and returns a cleaned magpie object.
}
\examples{
\dontrun{
a <- readSource(type = "OECD_Plastic", subtype = "Use_2019")
}
}
\seealso{
[readSource()]
}
\author{
Qianzhi Zhang
}
