% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOECDAffordableHousingDatabase.R
\name{readOECDAffordableHousingDatabase}
\alias{readOECDAffordableHousingDatabase}
\title{Read dwelling vacancy share from OECD Affordable Housing Database
OECD (2024), OECD Affordable Housing Database - indicator HM1.1. Housing stock and construction, https://oe.cd/ahd}
\usage{
readOECDAffordableHousingDatabase(subtype)
}
\arguments{
\item{subtype}{Specify column to be read. Can be "vacant", "holiday", "both"}
}
\description{
Read dwelling vacancy share from OECD Affordable Housing Database
OECD (2024), OECD Affordable Housing Database - indicator HM1.1. Housing stock and construction, https://oe.cd/ahd
}
\author{
Bennet Weiss
}
