% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOWID.R
\name{readOWID}
\alias{readOWID}
\title{Read data from Our World in Data.}
\usage{
readOWID(subtype)
}
\arguments{
\item{subtype}{Character string specifying the data and subfolder name to be read in. Supported:
- "shipping_costs": Get sea freight costs relative to 1930 in percent.}
}
\description{
Read data from Our World in Data.
}
\author{
Bennet Weiss
}
