% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPlasticsEurope.R
\name{readPlasticsEurope}
\alias{readPlasticsEurope}
\title{Read Plastics Europe Data into a magpie Object}
\usage{
readPlasticsEurope(subtype)
}
\arguments{
\item{subtype}{Character string specifying the dataset and scope.
Valid formats include:
- "Production_region" (plastic production in Mt, dimensions: region, year)
- "SectorShare_EU" (plastic demand shares (%) by use category in the EU, dimensions: category, year)
- "EoL_EU" (plastic waste treatment in the EU in Mt, dimensions: year, treatment)}
}
\value{
magpie object of the Plastics Europe data
}
\description{
This function reads Plastics Europe data on plastics production by region,
demand shares by use category and EoL treatment from Excel files,
based on a specified subtype, and returns a cleaned magpie object.
}
\examples{
\dontrun{
a <- readSource(type = "PlasticsEurope", subtype = "Production_region")
}
}
\seealso{
[readSource()]
}
\author{
Leonie Schweiger
}
