% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRASMI.R
\name{readRASMI}
\alias{readRASMI}
\title{Read material intensity (MI) from RASMI.
MI is a function of use/function type (RS, RM, NR) and structure type (S, T, C, M)}
\usage{
readRASMI(subtype = "concrete")
}
\arguments{
\item{subtype}{Material subtype. Possible values are "concrete", "steel", "plastics".}
}
\description{
Fishman, T., Mastrucci, A., Peled, Y. et al.
RASMI: Global ranges of building material intensities differentiated by region, structure, and function.
Sci Data 11, 418 (2024). https://doi.org/10.1038/s41597-024-03190-7
Data available on Zenodo: https://zenodo.org/records/10124952
Data available on Github: https://github.com/TomerFishman/MaterialIntensityEstimator
}
\author{
Bennet Weiss.
}
