% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRyberg.R
\name{readRyberg}
\alias{readRyberg}
\title{Read Ryberg et al. 2019 Global environmental losses of plastics across their value chains}
\usage{
readRyberg(subtype)
}
\arguments{
\item{subtype}{Character string specifying the dataset and scope
- "Sector_split"}
}
\value{
magpie object of the Ryberg et al. Plastic data
}
\description{
Read Ryberg et al. 2019 Global environmental losses of plastics across their value chains
}
\examples{
\dontrun{
a <- readSource(type = "Ryberg", subtype = "Sector_split")
}
}
\seealso{
[readSource()]
}
\author{
Leonie Schweiger
}
