% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUNCTAD.R
\name{readUNCTAD}
\alias{readUNCTAD}
\title{Read UNCTAD}
\usage{
readUNCTAD()
}
\value{
magpie object of the UNCTAD data
}
\description{
Read-in US_PlasticsTradebyPartner file from the
United Nations Conference on Trade and Development (UNCTAD).
Data is filtered for Trading Partner == World, i.e. all trading partners aggregated.
}
\seealso{
[readSource()]

\dontrun{
a <- readSource(type = "UNCTAD")
}
}
\author{
Qianzhi Zhang, Leonie Schweiger
}
