% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUNCTAD_PlasticsHSCodes.R
\name{readUNCTAD_PlasticsHSCodes}
\alias{readUNCTAD_PlasticsHSCodes}
\title{Read HS Comtrade product classification codes for plastics products from
UNCTAD}
\usage{
readUNCTAD_PlasticsHSCodes(subtype)
}
\arguments{
\item{subtype}{Character string specifying specifying the year of the HS (Harmonized System) revision of the data, available are
- 02
- 07
- 12
- 17
- 22}
}
\value{
magpie object of the UNEP NGP HS Codes
}
\description{
Read HS Comtrade product classification codes for plastics products from
UNCTAD
}
\examples{
\dontrun{
a <- readSource(type = "UNCTAD_PlasticsHSCodes")
}
}
\seealso{
[readSource()]
}
\author{
Leonie Schweiger
}
