% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUNEP_NGP.R
\name{readUNEP_NGP}
\alias{readUNEP_NGP}
\title{Read HS Comtrade product classification codes for plastics products and their plastics percentage from
Tool T1.4 of the UNEP National Guidance for Plastic Pollution Hotspotting and Shaping Action}
\usage{
readUNEP_NGP(subtype)
}
\arguments{
\item{subtype}{whether to return 4 digit or 6 digit codes or all
- k4
- k6
- all}
}
\value{
magpie object of the UNEP NGP HS Codes
}
\description{
Read HS Comtrade product classification codes for plastics products and their plastics percentage from
Tool T1.4 of the UNEP National Guidance for Plastic Pollution Hotspotting and Shaping Action
}
\examples{
\dontrun{
a <- readSource(type = "UNEP_NGP")
}
}
\seealso{
[readSource()]
}
\author{
Leonie Schweiger
}
