% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUS_EPA.R
\name{readUS_EPA}
\alias{readUS_EPA}
\title{Read US_EPA Data into a magpie Object}
\usage{
readUS_EPA()
}
\value{
magpie object of the US_EPA Data
}
\description{
This function reads US EPA data on plastics EoL treatment in the US.
Unit: thousands of U.S. tons
}
\examples{
\dontrun{
a <- readSource(type = "US_EPA")
}
}
\seealso{
[readSource()]
}
\author{
Leonie Schweiger
}
