% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUSGSDS140.R
\name{readUSGSDS140}
\alias{readUSGSDS140}
\title{Read cement statistic data sheet from USGS.
USGS DS140 combines historic data from USGS sources. Supply-Demand Statistics.
Imports pre-1950 include clinker. Export data include hydraulic cement and clinker.}
\usage{
readUSGSDS140(subtype)
}
\arguments{
\item{subtype}{Character string specifying the column to be read. Supported are:
- Imports
- Exports}
}
\description{
U.S. Geological Survey, 2020, Cement statistics, in Kelly, T.D., and Matos, G.R., comps.,
Historical statistics for mineral and material commodities in the United States:
U.S. Geological Survey Data Series 140
accessed 12.02.2025, at
https://www.usgs.gov/centers/national-minerals-information-center/
historical-statistics-mineral-and-material-commodities.
}
\author{
Bennet Weiss.
}
