% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWorldSteelParameters.R
\name{readWorldSteelParameters}
\alias{readWorldSteelParameters}
\title{TODOMERLIN: document}
\usage{
readWorldSteelParameters(subtype = "recoveryRate")
}
\arguments{
\item{subtype}{TODOMERLIN: document}
}
\description{
TODOMERLIN: document
}
\author{
Merlin Jo Hosak
}
