% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readXi2016.R
\name{readXi2016}
\alias{readXi2016}
\title{Read data from Supplementary Material in Xi et al. (2016)
"Substantial global carbon uptake by cement carbonation"
DOI: 10.1038/ngeo2840}
\usage{
readXi2016(subtype)
}
\arguments{
\item{subtype}{Variable to be read in.}
}
\description{
Read data from Supplementary Material in Xi et al. (2016)
"Substantial global carbon uptake by cement carbonation"
DOI: 10.1038/ngeo2840
}
\author{
Bennet Weiss
}
