% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolAddPlaceholder.R
\name{toolAddPlaceholder}
\alias{toolAddPlaceholder}
\title{Append placeholder categories to a magpie object.
Useful to extend data with "N/A" entries for categories that are not covered by the source data.}
\usage{
toolAddPlaceholder(x, names, fill = 1)
}
\arguments{
\item{x}{Magpie object}

\item{names}{Names of the new entries in dim 3. Subdimensions are separated by dots, e.g. "Ind.N/A".}

\item{fill}{Value the new entries are filled with.}
}
\description{
Append placeholder categories to a magpie object.
Useful to extend data with "N/A" entries for categories that are not covered by the source data.
}
\author{
Bennet Weiss
}
