% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolAggregateBilateralTrade.R
\name{toolAggregateBilateralTrade}
\alias{toolAggregateBilateralTrade}
\title{Aggregate bilateral trade to regional aggregation such that intra-regional trade is filtered out}
\usage{
toolAggregateBilateralTrade(x, rel, flow_label)
}
\arguments{
\item{x}{A magpie object with temporal dimension "t", spatial dimension "importer", and
further dimensions "exporter" and custom grouping variables.
The magpie object should actually contain two spatial dimensions "importer" and
"exporter", but since madrat does not support regional aggregation for two spatial
dimensions with iso countries well (as in trade data), we just use one spatial dimension
for the importer}

\item{flow_label}{The trade type: either "Exports" or "Imports"}

\item{ref}{A dataframe with the regional aggregation mapping with columns "country" and "region"}
}
\value{
An aggregated magpie object
}
\description{
Aggregate bilateral trade to regional aggregation such that intra-regional trade is filtered out
}
\author{
Leonie Schweiger
}
