% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolBackcastByReference2D.R
\name{toolBackcastByReference2D}
\alias{toolBackcastByReference2D}
\title{Backcast (or forecast) 2D magpie data by reference data}
\usage{
toolBackcastByReference2D(
  x,
  ref,
  doInterpolate = TRUE,
  maxN = 5,
  doForecast = FALSE,
  doMakeZeroNA = FALSE
)
}
\arguments{
\item{x}{A 2D magpie object (regions x years) to be backcasted/forecasted.}

\item{ref}{A 2D magpie object (regions x years) to be used as reference for
the backcasting/forecasting.}

\item{doInterpolate}{Logical, whether to interpolate missing values in x and ref
before calculating weights. Recommended to ensure more stable weight calculation.}

\item{maxN}{Maximum number of years to consider for weight calculation (default 5).}

\item{doForecast}{Logical, whether to do forecasting instead of backcasting.}

\item{doMakeZeroNA}{Logical, whether to convert 0 values in final output to NA.
As sometimes 0 is wrongly implicitly assumed.}
}
\description{
This function backcasts (or forecasts) missing values in a 2D magpie object x
based on reference data ref. The backcasting/forecasting is done by calculating
ratios of x to ref for shared years, computing weights for these ratios based
on their recency (linearly weighted), and applying these weights to scale ref
for missing years in x.
}
\author{
Merlin Jo Hosak
}
