% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCao2024.R
\name{toolCeCalculateMeans}
\alias{toolCeCalculateMeans}
\title{Calculate the data mean, do normalization check and reshape the data.}
\usage{
toolCeCalculateMeans(
  data,
  long_names,
  dim_members = NULL,
  dim = NULL,
  normalize = FALSE,
  tol = 0.03,
  fmean = FALSE,
  groups = NULL,
  warn = FALSE
)
}
\arguments{
\item{data}{original input from dataframe.}

\item{long_names}{long label of the first column of a variable.}

\item{dim_members}{names of the different elements of a dimension.}

\item{dim}{name of the dimension.}

\item{normalize}{Bool if normalization step should be performed.}

\item{tol}{Tolerance for check of normalization within rounding errors. Only performed if normalize = TRUE.}

\item{fmean}{Bool if fraction mean should be calculated.}

\item{groups}{optional vector (same length as long_names) giving group IDs.}

\item{warn}{Bool flag to activate warning if underlying data is not normalized.}
}
\description{
Calculate the data mean, do normalization check and reshape the data.
}
\author{
Bennet Weiss
}
