% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCao2024.R
\name{toolCeCalculateMeanSingle}
\alias{toolCeCalculateMeanSingle}
\title{Function to calculate the mean by extracting the distribution function.}
\usage{
toolCeCalculateMeanSingle(data, start_column_name, fmean = FALSE)
}
\arguments{
\item{data}{original input from dataframe.}

\item{start_column_name}{label of first column of variable.}

\item{fmean}{Bool if fraction mean should be calculated.}
}
\description{
Function to calculate the mean by extracting the distribution function.
}
\author{
Bennet Weiss
}
