% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCeCumulativeCementProduction.R
\name{toolCeCumulativeCementProduction}
\alias{toolCeCumulativeCementProduction}
\title{Returns cumulative cement production casted to given dimensions.
Can be used as weight in cement-related aggregation without time component.
Assumes time dimension is in 2nd position.}
\usage{
toolCeCumulativeCementProduction(castto = NULL)
}
\arguments{
\item{castto}{magpie object that provides the desired dimensions.
Defaults to NULL, in which case no dimension casting is performed.}
}
\description{
Returns cumulative cement production casted to given dimensions.
Can be used as weight in cement-related aggregation without time component.
Assumes time dimension is in 2nd position.
}
\author{
Bennet Weiss
}
