% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCleanSteelRegions.R
\name{toolCleanSteelRegions}
\alias{toolCleanSteelRegions}
\title{Clean up region names in steel sources}
\usage{
toolCleanSteelRegions(df)
}
\arguments{
\item{df}{a data frame with steel data consisting of three columns 'country_name',
'period' and 'value'}
}
\description{
Clean up region names for a given data frame with steel data
read in from in \link{readWorldSteelDigitised} and \link{readWorldSteelDatabase}

Applies data cleansing, removes ignored countries/regions and converts country names to ISO3
country codes.

This operation would normally happen on magclass object in the convert function,
but as the cleansing can introduce duplicates for the given data, we resolve this on
the data frames in the respective read functions before converting to a MagPIE
object (MagPIE does not handle duplicate regions in first dimension well)
}
\author{
Falk Benke
}
