% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGEM.R
\name{toolExtractNStories}
\alias{toolExtractNStories}
\title{Extracts the int number of stories for a taxonomy string as used in GEM.}
\usage{
toolExtractNStories(taxonomy)
}
\arguments{
\item{taxonomy}{string or list of strings to check.}
}
\description{
Extracts the int number of stories for a taxonomy string as used in GEM.
}
\author{
Bennet Weiss
}
