% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCao2024.R
\name{toolFmeanTruncWeibull}
\alias{toolFmeanTruncWeibull}
\title{Conditional harmonic mean (1 / E[1/X]) for truncated Weibull on [a,b].
Parameters: (scale, shape, min, max). Requires (shape>1 or a>0). Diverges if a=0 & shape<=1.
Returns 1 / E[1/X | a<=X<=b].}
\usage{
toolFmeanTruncWeibull(parameters)
}
\arguments{
\item{parameters}{Array of the four parameters (scale, shape, min, max) of a Weibull distribution}
}
\description{
Conditional harmonic mean (1 / E[1/X]) for truncated Weibull on [a,b].
Parameters: (scale, shape, min, max). Requires (shape>1 or a>0). Diverges if a=0 & shape<=1.
Returns 1 / E[1/X | a<=X<=b].
}
\author{
Bennet Weiss
}
