% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCao2024.R
\name{toolFmeanUniform}
\alias{toolFmeanUniform}
\title{Conditional harmonic mean (1 / E[1/X]) for Uniform(a,b).
Parameters layout kept: col1 = b (max), col2 = a (min). Requires 0 < a < b.
Returns (b - a)/(log b - log a).}
\usage{
toolFmeanUniform(parameters)
}
\arguments{
\item{parameters}{Array of the two parameters (max, min) of a uniform distribution}
}
\description{
Conditional harmonic mean (1 / E[1/X]) for Uniform(a,b).
Parameters layout kept: col1 = b (max), col2 = a (min). Requires 0 < a < b.
Returns (b - a)/(log b - log a).
}
\author{
Bennet Weiss
}
