% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGlobalExposureModel.R
\name{toolInferResBuildingType}
\alias{toolInferResBuildingType}
\title{Categorizes res buildings into single family (RS) or multi family (RM) homes based on their number of stories.}
\usage{
toolInferResBuildingType(data)
}
\arguments{
\item{data}{Dataframe that contains a column OCCUPANCY with building function types
and a column TAXONOMY with taxonomy strings.}
}
\description{
Categorizes res buildings into single family (RS) or multi family (RM) homes based on their number of stories.
}
\author{
Bennet Weiss
}
