% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolInterpolate2D.R
\name{toolInterpolate2D}
\alias{toolInterpolate2D}
\title{Interpolate missing values in a 2D magpie object}
\usage{
toolInterpolate2D(x, method = "linear", ...)
}
\arguments{
\item{x}{A 2D magpie object (regions x years) with missing values to interpolate}

\item{method}{3 interpolation methods are supported:
'linear' (default), 'spline', and 'constant' (last observation carried forward).}

\item{...}{Additional arguments passed to interpolation functions}
}
\value{
A 2D magpie object with interpolated values
}
\description{
This function interpolates missing values (NAs) in a 2D magpie
object along the time dimension (years).
}
\author{
Merlin Jo Hosak
}
