% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolNAHandling.R
\name{toolMaskNACountries}
\alias{toolMaskNACountries}
\title{Create mask for countries that have less than n_valid values which are not NA.}
\usage{
toolMaskNACountries(x, n_valid = 5)
}
\arguments{
\item{x}{Magpie object}

\item{n_valid}{Threshold minimum number of valid (not NA) values.}
}
\description{
Create mask for countries that have less than n_valid values which are not NA.
}
\author{
Bennet Weiss
}
