% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCao2024.R
\name{toolMeanTriangular}
\alias{toolMeanTriangular}
\title{Calculate the mean based on parameters from a triangular distribution.}
\usage{
toolMeanTriangular(parameters)
}
\arguments{
\item{parameters}{Array of the three parameters (mode, max, min) of a triangular distribution}
}
\description{
Calculate the mean based on parameters from a triangular distribution.
}
\author{
Bennet Weiss
}
