% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCao2024.R
\name{toolMeanTruncNorm}
\alias{toolMeanTruncNorm}
\title{Calculate the mean based on parameters from a truncated normal distribution.}
\usage{
toolMeanTruncNorm(parameters)
}
\arguments{
\item{parameters}{Array of the four parameters (mean, std, min, max) of a truncated normal distribution}
}
\description{
Calculate the mean based on parameters from a truncated normal distribution.
}
\author{
Bennet Weiss
}
