% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCao2024.R
\name{toolMeanTruncWeibull}
\alias{toolMeanTruncWeibull}
\title{Calculate the mean based on parameters from a truncated Weibull distribution.}
\usage{
toolMeanTruncWeibull(parameters)
}
\arguments{
\item{parameters}{Array of the four parameters (scale, shape, min, max) of a Weibull distribution}
}
\description{
Calculate the mean based on parameters from a truncated Weibull distribution.
}
\author{
Bennet Weiss
}
