% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCao2024.R
\name{toolMeanUniform}
\alias{toolMeanUniform}
\title{Calculate the mean based on parameters from a uniform distribution.}
\usage{
toolMeanUniform(parameters)
}
\arguments{
\item{parameters}{Array of the two parameters (max, min) of a uniform distribution}
}
\description{
Calculate the mean based on parameters from a uniform distribution.
}
\author{
Bennet Weiss
}
