% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolMerge2D.R
\name{toolMerge2D}
\alias{toolMerge2D}
\title{Merge two 2D magpie objects}
\usage{
toolMerge2D(a, b, fill = NA)
}
\arguments{
\item{a}{First magpie object to merge}

\item{b}{Second magpie object to merge}

\item{fill}{Value to use for cells not present in either input (default: NA)}
}
\value{
A 2D magpie object containing the union of regions and years from both inputs
}
\description{
Perform an outer merge on two 2D magpie objects (regions x years).
}
\author{
Merlin Jo Hosak
}
