% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBACI.R
\name{calcBACI}
\alias{calcBACI}
\title{Get BACI Trade data for specific product groups}
\usage{
calcBACI(subtype, HS = "02", category)
}
\arguments{
\item{subtype}{Character string specifying the scope
- "plastics_UNCTAD": plastics trade data for HS codes from UNCTAD classification into
  primary, intermediate, manufactured, final and waste plastics
- "plastics_UNEP": plastics trade data for HS codes from UNEP NGP
  (estimated plastics mass based on estimated plastics percentages in goods)}

\item{HS}{Character string specifying the year of the HS (Harmonized System) revision of the data
- 92
- 02
- 17
- 22}

\item{category}{Character string specifying the stage of trade, valid parameters depend on the subtype
for plastics_UNCTAD:
- Plastics in primary forms
- Intermediate forms of plastic
- Intermediate manufactured plastic goods
- Final manufactured plastics goods
- Plastic waste
for plastics_UNEP:
- Primary
- Application
- Waste}
}
\value{
magpie object of the BACI trade data
}
\description{
Get BACI Trade data for specific product groups
}
\examples{
\dontrun{
a <- calcOutput(type = "BACI", subtype = "plastics_UNCTAD", HS = "02", category = "Plastics in primary forms")
}
}
\seealso{
[calcOutput()]
}
\author{
Leonie Schweiger
}
