% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeBuildingsMI.R
\name{calcCeBuildingsMI}
\alias{calcCeBuildingsMI}
\title{Calculates global cement production as from Andrew's 2019 paper.}
\usage{
calcCeBuildingsMI(subtype = "concrete")
}
\arguments{
\item{subtype}{Material subtype. Possible values are "concrete", "steel", "plastics".}
}
\description{
Calculates global cement production as from Andrew's 2019 paper.
}
\author{
Bennet Weiss
}
