% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeBuiltLifespan.R
\name{calcCeBuiltLifespan}
\alias{calcCeBuiltLifespan}
\title{Calculates the lifetimes of residential and non-residential buildings, as well as of civil engineering.}
\usage{
calcCeBuiltLifespan()
}
\description{
Calculates the lifetimes of residential and non-residential buildings, as well as of civil engineering.
}
\author{
Bennet Weiss
}
