% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeFloorspace.R
\name{calcCeFloorspace}
\alias{calcCeFloorspace}
\title{Calculates the floorspace of residential (Res) and commercial (Com) buildings.
Based on floorspace data from EDGE-B.
Corrected by region-specific calibration using EUBUCCO and GEM if correct = TRUE.}
\usage{
calcCeFloorspace(scenario = "SSP2", correct = TRUE)
}
\arguments{
\item{scenario}{EDGE-B scenario, one of "SSP1", "SSP2", "SSP3", "SSP4", "SSP5".}

\item{correct}{Bool, whether to apply the correction factor (default: TRUE).}
}
\description{
Calculates the floorspace of residential (Res) and commercial (Com) buildings.
Based on floorspace data from EDGE-B.
Corrected by region-specific calibration using EUBUCCO and GEM if correct = TRUE.
}
\author{
Bennet Weiss
}
