% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeFloorspaceEDGEB.R
\name{calcCeFloorspaceEDGEB}
\alias{calcCeFloorspaceEDGEB}
\title{Calculates the buildings floor area from EDGE-B}
\usage{
calcCeFloorspaceEDGEB(scenario = "SSP2")
}
\arguments{
\item{scenario}{EDGE-B scenario, one of "SSP1", "SSP2", "SSP3", "SSP4", "SSP5"}
}
\description{
Calculates the buildings floor area from EDGE-B
}
\author{
Bennet Weiss
}
