% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeTrade.R
\name{calcCeTrade}
\alias{calcCeTrade}
\title{Split World Steel trade into bilateral trade by using proxy bilateral BACI trade data
and aggregate via a custom aggregation function to filter out intra-regional trade
(since we don't have a list with all HS codes and corresponding steel shares,
BACI trade data cannot be used directly, since it does not cover the full scope)}
\usage{
calcCeTrade(
  subtype,
  category,
  HS = "92",
  include_intra_regional = FALSE,
  target_years = NULL
)
}
\arguments{
\item{subtype}{Character string specifying the scope
- Imports
- Exports}

\item{category}{Character string specifying the stage of trade
- cement
- clinker}

\item{HS}{Character string specifying the year of the HS (Harmonized System) revision of the data
- 92
- 02
- 17
- 22}

\item{include_intra_regional}{bool if intra-regional trade should be included}

\item{target_years}{integer vector of target years for the output data.
If NULL, all years from reference (cement production) are included.
Note: the 'years' argument in calcOutput does not work properly for this function,
so 'target years' should be set here instead.}
}
\value{
magpie object of the aggregated trade data
}
\description{
Split World Steel trade into bilateral trade by using proxy bilateral BACI trade data
and aggregate via a custom aggregation function to filter out intra-regional trade
(since we don't have a list with all HS codes and corresponding steel shares,
BACI trade data cannot be used directly, since it does not cover the full scope)
}
\examples{
\dontrun{
a <- calcOutput(
  type = "CeTrade", subtype = "Imports",
  category = "cement", HS = "92"
)
}
}
\seealso{
[calcOutput()]
}
\author{
Bennet Weiss, Leonie Schweiger
}
