% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCoPopulation1900To2150.R
\name{calcCoPopulation1900To2150}
\alias{calcCoPopulation1900To2150}
\title{Get population from 1900-2100}
\usage{
calcCoPopulation1900To2150(scenario = "SSP2")
}
\arguments{
\item{scenario}{Scenario to use for future population data (default: SSP2).}
}
\value{
List with Magpie object of population and metadata in calcOutput
format.
}
\description{
Calc population from 1900-2100 yearly for the REMIND-MFA format on a country
level. Can be aggregated to regions via calcOutput aggregate parameter.
Uses \link{readGapminder} and \link[mrdrivers]{readUN_PopDiv}
datasets for historical population data as well as
\link[mrdrivers]{calcPopulation} from
mrdrivers for current and future population data according to a specific
scenario (see \code{vignette("scenarios")} for more information).
}
\author{
Merlin Jo Hosak
}
