% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlCarbonContent.R
\name{calcPlCarbonContent}
\alias{calcPlCarbonContent}
\title{Returns the carbon contents of different plastic types/polymers}
\usage{
calcPlCarbonContent()
}
\description{
Returns the carbon contents of different plastic types/polymers
}
\author{
Leonie Schweiger
}
