% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlMechLoss.R
\name{calcPlMechLoss}
\alias{calcPlMechLoss}
\title{Calculate Country-Level Mechanical Recycling Loss Trajectories}
\usage{
calcPlMechLoss()
}
\description{
Generate time series of mechanical recycling loss trajectories by sector and region,
then aggregate to countries for 1990–2100.
}
\author{
Qianzhi Zhang
}
