% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlMechReYield.R
\name{calcPlMechReYield}
\alias{calcPlMechReYield}
\title{Calculate Country-Level Mechanical Recycling Yield Trajectories}
\usage{
calcPlMechReYield()
}
\description{
Generate time series of mechanical recycling yield (efficiency) trajectories by sector and region,
then aggregate to countries for 1990-2100.
}
\author{
Qianzhi Zhang
}
