% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlOECD_MGshare.R
\name{calcPlOECD_MGshare}
\alias{calcPlOECD_MGshare}
\title{Calculate Country-Level Material Good Shares}
\usage{
calcPlOECD_MGshare()
}
\description{
Compute material share of different goods from 2019 OECD use data,
aggregate regional shares to country level for 2019.
}
\author{
Qianzhi Zhang
}
