% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlPottinger.R
\name{calcPlPottinger}
\alias{calcPlPottinger}
\title{Pottinger 2024 plastics material flows in IAMC variables}
\usage{
calcPlPottinger()
}
\value{
List with a MagPIE object of plastics flows (Mt/yr) in IAMC variables
and metadata in calcOutput format.
}
\description{
Return the Business-as-Usual plastics material-flow data from Pottinger et al.
(2024) at ISO3 country level, with the source flow columns renamed to IAMC
variables under the \code{Materials|Plastics|...} hierarchy. All flows are in
Mt/yr and are absolute quantities, so \code{weight = NULL} (they are summed
when aggregated to regions).
}
\seealso{
\code{\link{readPottinger2024}}
}
\author{
Leonie Schweiger
}
