% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlSyntheticFibre.R
\name{calcPlSyntheticFibre}
\alias{calcPlSyntheticFibre}
\title{Calculate synthetic fibre production by country}
\usage{
calcPlSyntheticFibre()
}
\value{
A list in \code{\link[madrat]{calcOutput}} format with total synthetic
  fibre production by country and year.
}
\description{
Combines the global synthetic fibre production time series (Polyester,
Polyamide (nylon), Acrylic, Polypropylene and Elastane) from Textile Exchange
with the estimated country-level production shares based on Credence Research
market data on monoethylene glycol (MEG) to yield a total synthetic
fibre production by country and year.
}
\examples{
\dontrun{
a <- calcOutput("PlSyntheticFibre")
}
}
\seealso{
[readTextileExchange()]
}
\author{
Leonie Schweiger
}
