% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlTrade.R
\name{calcPlTrade}
\alias{calcPlTrade}
\title{Calculate Country-Level Plastics Trade for Various Categories}
\usage{
calcPlTrade(
  category,
  flow_label = c("Exports", "Imports"),
  data_source = c("UNCTAD", "BACI_UNCTAD", "BACI_UNEP"),
  HS = "92",
  target_years = NULL
)
}
\arguments{
\item{category}{Character; product category:
\itemize{
  \item "Final"        - Final plastics
  \item "Primary"      - Primary plastics
  \item "Intermediate" - Intermediate forms of plastic
  \item "Manufactured" - Intermediate manufactured plastic goods
  \item "Application"  - Plastic goods
  \item "Waste"        - Plastic waste
}}

\item{flow_label}{Character; trade flow:
\itemize{
  \item "Exports" - Exports
  \item "Imports" - Imports
}}

\item{data_source}{Character; data source:
\itemize{
  \item "UNCTAD" - UNCTAD (trade flows by time and region)
  \item "BACI_UNCTAD" - BACI with UNCTAD HS codes (trade flows by time and region)
  \item "BACI_UNEP" - BACI with UNEP HS codes (trade flows by time, region, sector and polymer)
}}

\item{HS}{Character string specifying the year of the HS (Harmonized System) revision of the data
- 92
- 02
- 17
- 22}

\item{target_years}{integer vector of target years for the output data.
If NULL, all years from reference (plastics production) are included.
Note: the 'years' argument in calcOutput does not work properly for this function,
so 'target years' should be set here instead.}
}
\description{
Reads plastics trade data and backcasts data to 1950 to fill missing years.
Note that as the source BACI has bilateral trade data that allows to filter out
intraregional trade, aggregation is done by a custom aggregation function for this source.
}
\author{
Qianzhi Zhang, Leonie Schweiger
}
