% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlTrade.R
\name{calcPlTrade}
\alias{calcPlTrade}
\title{Calculate Country-Level Plastics Trade for Various Categories}
\usage{
calcPlTrade(
  category = c("Final", "Primary", "Intermediate", "Manufactured"),
  flow_label = c("Exports", "Imports")
)
}
\arguments{
\item{category}{Character; product category:
\itemize{
  \item "Final"        - Final plastics
  \item "Primary"      - Primary plastics
  \item "Intermediate" - Intermediate forms of plastic
  \item "Manufactured" - Intermediate manufactured plastic goods
}}

\item{flow_label}{Character; trade flow:
\itemize{
  \item "Exports" - Exports
  \item "Imports" - Imports
}}
}
\description{
Reads UNCTAD plastics trade (exports or imports) data at regional level,
backcasts data to 1950 to fill missing years 1950-2004 years,
and aggregates to country level.
}
\author{
Qianzhi Zhang
}
