% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlTrade.R
\name{calcPlTrade}
\alias{calcPlTrade}
\title{Calculate Country-Level Plastics Trade for Various Categories}
\usage{
calcPlTrade(
  category,
  flow_label = c("Exports", "Imports"),
  data_source = c("UNCTAD", "BACI")
)
}
\arguments{
\item{category}{Character; product category:
\itemize{
  \item "Final"        - Final plastics
  \item "Primary"      - Primary plastics
  \item "Intermediate" - Intermediate forms of plastic
  \item "Manufactured" - Intermediate manufactured plastic goods
  \item "Application"  - Plastic goods
  \item "Waste"        - Plastic waste
}}

\item{flow_label}{Character; trade flow:
\itemize{
  \item "Exports" - Exports
  \item "Imports" - Imports
}}

\item{data_source}{Character; data source:
\itemize{
  \item "UNCTAD" - UNCTAD (trade flows by time and region)
  \item "BACI" - BACI (trade flows by time, region, sector and polymer)
}}
}
\description{
Reads plastics trade (exports or imports) data at regional level,
and backcasts data to 1950 to fill missing years.
Note that aggregation to regions is done in the calc functions called by this function,
as the source BACI has bilateral trade data that allows to filter out intraregional trade
by a custom aggregation function.
Therefore, we operate already at the regional aggregation level here (isocountries=FALSE).
}
\author{
Qianzhi Zhang
}
