% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlTrade.R
\name{calcPlTrade}
\alias{calcPlTrade}
\title{Calculate Country-Level Plastics Trade for Various Categories}
\usage{
calcPlTrade(
  category = c("final", "primary", "intermediate", "manufactured"),
  flow_label = c("Exports", "Imports")
)
}
\arguments{
\item{category}{Character; product category:
\itemize{
  \item "final"        - Final plastics
  \item "primary"      - Primary plastics
  \item "intermediate" - Intermediate forms of plastic
  \item "manufactured" - Intermediate manufactured plastic goods
}}

\item{flow_label}{Character; trade flow:
\itemize{
  \item "Exports" - Exports
  \item "Imports" - Imports
}}
}
\description{
Reads UNCTAD plastics trade (exports or imports) data at regional level,
fills 1990-2004 years with data of 2005, and aggregates to country level.
}
\author{
Qianzhi Zhang
}
