% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlTrade.R
\name{calcPlTrade}
\alias{calcPlTrade}
\title{Calculate Country-Level Plastics Trade for Various Categories}
\usage{
calcPlTrade(
  category,
  flow_label = c("Exports", "Imports"),
  data_source = c("UNCTAD", "BACI_UNCTAD", "BACI_UNEP")
)
}
\arguments{
\item{category}{Character; product category:
\itemize{
  \item "Final"        - Final plastics
  \item "Primary"      - Primary plastics
  \item "Intermediate" - Intermediate forms of plastic
  \item "Manufactured" - Intermediate manufactured plastic goods
  \item "Application"  - Plastic goods
  \item "Waste"        - Plastic waste
}}

\item{flow_label}{Character; trade flow:
\itemize{
  \item "Exports" - Exports
  \item "Imports" - Imports
}}

\item{data_source}{Character; data source:
\itemize{
  \item "UNCTAD" - UNCTAD (trade flows by time and region)
  \item "BACI_UNCTAD" - BACI with UNCTAD HS codes (trade flows by time and region)
  \item "BACI_UNEP" - BACI with UNEP HS codes (trade flows by time, region, sector and polymer)
}}
}
\description{
Reads plastics trade data and backcasts data to 1950 to fill missing years.
Note that as the source BACI has bilateral trade data that allows to filter out
intraregional trade, aggregation is done by a custom aggregation function for this source.
}
\author{
Qianzhi Zhang, Leonie Schweiger
}
