% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStCullenFabricationYield.R
\name{calcStCullenFabricationYield}
\alias{calcStCullenFabricationYield}
\title{TODOMERLIN: document}
\usage{
calcStCullenFabricationYield()
}
\description{
TODOMERLIN: document
}
\author{
Merlin Jo Hosak
}
