% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStIndirectTradeShares.R
\name{calcStIndirectTradeShares}
\alias{calcStIndirectTradeShares}
\title{Calculate indirect trade shares from World Steel Digitised data on indirect
trade in 2013.}
\usage{
calcStIndirectTradeShares()
}
\description{
Calculate indirect trade shares from World Steel Digitised data on indirect
trade in 2013.
}
\author{
Falk Benke
}
