% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStLifetimes.R
\name{calcStLifetimes}
\alias{calcStLifetimes}
\title{Calc Steel Lifetimes}
\usage{
calcStLifetimes(subtype)
}
\arguments{
\item{subtype}{TODOMERLIN: document}
}
\description{
Function to load steel lifetimes data in years.
}
\author{
Merlin Jo Hosak
}
