% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStLifetimes.R
\name{calcStLifetimes}
\alias{calcStLifetimes}
\title{TODOMERLIN: document}
\usage{
calcStLifetimes(subtype)
}
\arguments{
\item{subtype}{TODOMERLIN: document}
}
\description{
TODOMERLIN: document
}
\author{
Merlin Jo Hosak
}
