% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStRecoveryRate.R
\name{calcStRecoveryRate}
\alias{calcStRecoveryRate}
\title{TODOMERLIN: document}
\usage{
calcStRecoveryRate(subtype)
}
\arguments{
\item{subtype}{TODOMERLIN: document}
}
\description{
TODOMERLIN: document
}
\author{
Merlin Jo Hosak
}
