% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStSectorSplits.R
\name{calcStSectorSplits}
\alias{calcStSectorSplits}
\title{Calc steel sector splits}
\usage{
calcStSectorSplits(subtype)
}
\arguments{
\item{subtype}{Income bin; either "high" or "low"}
}
\description{
Function to load steel sector splits data in the four main end-use sectors
(construction, machinery, products, transport). Currently summarized data
from Pauliuk et al. (2013) is used that is only available for the US in
2004, India in 1995-1999 and the UK in 1960-65, 1970, 1974-79. These are
based on the American Iron and Steel Institute, Spark Steel & Economy
Research, the Iron and Steel Statistics Bureau and Dahlström et al. (2004)
(see Metadata in File). See \link{readPauliuk2013} for
preprocessing.
}
\author{
Merlin Jo Hosak
}
