% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStSectorSplits.R
\name{calcStSectorSplits}
\alias{calcStSectorSplits}
\title{TODOMERLIN: document}
\usage{
calcStSectorSplits(subtype)
}
\arguments{
\item{subtype}{TODOMERLIN: document}
}
\description{
TODOMERLIN: document
}
\author{
Merlin Jo Hosak
}
