% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGapminder.R
\name{convertGapminder}
\alias{convertGapminder}
\title{Convert Gapminder data to Magpie format}
\usage{
convertGapminder(x, subtype = "population")
}
\description{
Convert Gapminder data to Magpie format
}
\author{
Merlin Jo Hosak
}
