% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertOECD_GDP.R
\name{convertOECD_GDP}
\alias{convertOECD_GDP}
\title{Convert OECD GDP}
\usage{
convertOECD_GDP(x)
}
\arguments{
\item{x}{MagPIE object}
}
\description{
Convert OECD GDP per capita data from 1500-2016
}
\author{
Merlin Jo Hosak
}
