% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertOECD_GDP.R
\name{convertOECD_GDP}
\alias{convertOECD_GDP}
\title{Conver OECD GDP}
\usage{
convertOECD_GDP(x, subtype = "gdpPC")
}
\arguments{
\item{x}{Magpie object as read by readOECD_GDP}

\item{subtype}{Specific dataset used by that source}
}
\description{
Convert OECD GDP per capita data from 1500-2016
}
\author{
Merlin Jo Hosak
}
