% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPlasticsEurope.R
\name{convertPlasticsEurope}
\alias{convertPlasticsEurope}
\title{Convert Plastics Europe production data to ISO country level}
\usage{
convertPlasticsEurope(x)
}
\arguments{
\item{x}{MagPIE object containing Plastics Europe production data at regional resolution.}
}
\value{
MagPIE object of the Plastics Europe production data disaggregated to country level.
}
\description{
Convert Plastics Europe production data to ISO country level
}
\examples{
\dontrun{
a <- convertPlasticsEurope(x)
}
}
\author{
Leonie Schweiger
}
