% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertWorldSteelDatabase.R
\name{convertWorldSteelDatabase}
\alias{convertWorldSteelDatabase}
\title{Convert data from WorldSteelAssociation Database}
\usage{
convertWorldSteelDatabase(x, subtype = "production")
}
\arguments{
\item{x}{Magpie object}
}
\description{
Convert data from WorldSteelAssociation Database
}
\author{
Merlin Jo Hosak
}
