% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertWorldSteelDatabase.R
\name{convertWorldSteelDatabase}
\alias{convertWorldSteelDatabase}
\title{Convert World Steel Database}
\usage{
convertWorldSteelDatabase(x, subtype)
}
\arguments{
\item{x}{Magpie object}

\item{subtype}{must be one of "production", "bofProduction", "eafProduction",
"imports", "exports", "scrapImports",  "scrapExports", "indirectImports",
"indirectExports", "pigIronProduction", "pigIronImports", "pigIronExports",
"driProduction", "driImports", "driExports"}
}
\description{
Convert data from WorldSteelAssociation Database
}
\author{
Merlin Jo Hosak
}
