% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertWorldSteelDigitised.R
\name{convertWorldSteelDigitised}
\alias{convertWorldSteelDigitised}
\title{Convert World Steel Digitised}
\usage{
convertWorldSteelDigitised(x, subtype)
}
\arguments{
\item{x}{MagPIE object}

\item{subtype}{Available subtypes are:
'worldProduction', 'production', 'productionByProcess', 'imports', 'exports',
'scrapImports', 'scrapExports', 'scrapConsumption', historicScrapShare',
'worldScrapConsumption', 'indirectTrade'}
}
\description{
Convert data World Steel Association digitised 1978-2022 yearbooks.
}
\author{
Merlin Jo Hosak
}
