% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullMFA.R
\name{fullMFA}
\alias{fullMFA}
\title{Load all MFA data}
\usage{
fullMFA(
  rev = 0,
  dev = "",
  scenario = "SSP2",
  gdp_per_capita = FALSE,
  run_sections = NULL
)
}
\arguments{
\item{run_sections}{Character vector selecting which parts to run.
Allowed: c("drivers","steel","cement"). NULL (default) runs all.}
}
\description{
Function that produces the complete regional data sets required for the
MFA model.
}
\examples{
\dontrun{
retrieveData('MFA')
fullMFA()
}

}
\seealso{
\code{\link[madrat]{readSource}}, \code{\link[madrat]{getCalculations}},
\code{\link[madrat]{calcOutput}}, \code{\link[mrmfa]{calcSteelProduction}}
}
\author{
Merlin HOSAK

Bennet Weiss
}
