% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullMFA.R
\name{fullMFA}
\alias{fullMFA}
\title{Load all MFA data}
\usage{
fullMFA(
  rev = 0,
  dev = "",
  scenario = "SSP2",
  gdpPerCapita = TRUE,
  runSections = NULL,
  end_future = 2100
)
}
\arguments{
\item{rev}{Revision number for the data version}

\item{dev}{Development version string}

\item{scenario}{SSP scenario used for population and GDP drivers}

\item{gdpPerCapita}{bool if GDP driver should be returned as per capita values}

\item{runSections}{Character vector selecting which parts to run.}

\item{end_future}{End year for future data (default: 2100)
Allowed values (see validSections): c("drivers", "steel", "cement", "plastic"). NULL (default) runs all.}

\item{start_historic}{Start year for historic data (default: 1900)}

\item{end_historic}{End year for historic data (default: 2023)}
}
\description{
Function that produces the complete regional data sets required for the
MFA model.
}
\examples{
\dontrun{
retrieveData("MFA")
fullMFA()
}

}
\seealso{
\code{\link[madrat]{readSource}}, \code{\link[madrat]{getCalculations}},
\code{\link[madrat]{calcOutput}}
}
\author{
Merlin HOSAK

Bennet Weiss
}
