% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullMFA.R
\name{fullMFA}
\alias{fullMFA}
\title{Load all MFA data}
\usage{
fullMFA(
  rev = 0,
  dev = "",
  scenario = "SSP2",
  gdp_per_capita = FALSE,
  run_sections = NULL
)
}
\arguments{
\item{rev}{Revision number for the data version}

\item{dev}{Development version string}

\item{scenario}{SSP scenario used for population and GDP drivers}

\item{gdp_per_capita}{bool if GDP driver should be returned as per capita values}

\item{run_sections}{Character vector selecting which parts to run.
Allowed values (see validSections): c("drivers", "steel", "cement", "plastic"). NULL (default) runs all.}
}
\description{
Function that produces the complete regional data sets required for the
MFA model.
}
\examples{
\dontrun{
retrieveData("MFA")
fullMFA()
}

}
\seealso{
\code{\link[madrat:readSource]{readSource}}, \code{\link[madrat:getCalculations]{getCalculations}},
\code{\link[madrat:calcOutput]{calcOutput}}
}
\author{
Merlin HOSAK

Bennet Weiss
}
