% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeFloorspaceCorrectionFactor.R
\name{MplotMulti}
\alias{MplotMulti}
\title{mplot_multi}
\usage{
MplotMulti(
  px,
  global = TRUE,
  total = FALSE,
  title = NULL,
  xlab = NULL,
  ylab = NULL,
  filename = NULL,
  nrow = NULL,
  ncol = NULL,
  ignore_na = TRUE,
  show_shares = TRUE
)
}
\arguments{
\item{px}{The magpie object or a named list of magpie objects to be visualized.}

\item{global}{Whether data should be aggregated over regions to global values.}

\item{total}{Whether the total of all data values should also be visualized.}

\item{title}{A string to be used as the plot's main title. Defaults to NULL.}

\item{nrow}{Number of rows for the facet layout (passed to facet_wrap).}

\item{ncol}{Number of columns for the facet layout (passed to facet_wrap).}

\item{ignore_na}{Logical indicating whether NA, NaN, Inf, and -Inf values should be ignored (removed) before plotting. Defaults to TRUE.}

\item{show_shares}{Logical indicating whether stacked bars should display percentage shares for each category. Defaults to FALSE.}
}
\description{
Get a quick visualization of the content of most magpie objects.
Based on magclass::mplot. Now supports plotting a named list of magclass objects for comparison.
}
\author{
Bennet Weiss, Pascal Sauer (mplot), Patrick Rein (mplot)
}
