% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrmfa-package.R
\docType{package}
\name{mrmfa-package}
\alias{mrmfa}
\alias{mrmfa-package}
\title{mrmfa: Input data generation for the REMIND MFA}
\description{
The mrmfa packages contains data preprocessing for the REMIND MFA.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/mrmfa}
}

}
\author{
\strong{Maintainer}: Jakob Dürrwächter \email{jakobdu@pik-potsdam.de}

Authors:
\itemize{
  \item Bennet Weiss \email{bennet.weiss@pik-potsdam.de}
  \item Leonie Schweiger \email{leonie.schweiger@pik-potsdam.de}
}

}
\keyword{internal}
