% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrmfa-package.R
\docType{package}
\name{mrmfa-package}
\alias{mrmfa}
\alias{mrmfa-package}
\title{mrmfa: Add an adequate Title}
\description{
Add an adequate Description, formulated as a complete sentence.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/mrmfa}
}

}
\author{
\strong{Maintainer}: Jakob Dürrwächter \email{jakobdu@pik-potsdam.de}

}
\keyword{internal}
