% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCeFloorspaceCorrectionFactor.R
\name{plot_floor_area_comparison}
\alias{plot_floor_area_comparison}
\title{Plots a comparison of floor area data from EDGE-B, EUBUCCO and GEM.}
\usage{
plot_floor_area_comparison(
  edgeb_floor_area,
  eubucco_floor_area,
  gem_floor_area,
  ghsoobat_floor_area
)
}
\description{
Plots a comparison of floor area data from EDGE-B, EUBUCCO and GEM.
}
\author{
Bennet Weiss
}
