% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBACI.R
\name{readBACI}
\alias{readBACI}
\title{Read BACI Trade data}
\usage{
readBACI(subtype, HS = "02")
}
\arguments{
\item{subtype}{Character string specifying the scope
- "plastics"}

\item{HS}{Character string specifying the year of the HS (Harmonized System) revision of the data
- 92
- 02
- 22}
}
\value{
magpie object of the BACI trade data
}
\description{
Read BACI Trade data
}
\examples{
\dontrun{
a <- readSource(type = "BACI", subtype = "plastics")
}
}
\seealso{
[readSource()]
}
\author{
Leonie Schweiger
}
