% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCooper2014.R
\name{readCooper2014}
\alias{readCooper2014}
\title{TODOMERLIN: document}
\usage{
readCooper2014(subtype = "lifetimes")
}
\arguments{
\item{subtype}{TODOMERLIN: document}
}
\description{
TODOMERLIN: document
}
\author{
Merlin Jo Hosak
}
