% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCullen2012.R
\name{readCullen2012}
\alias{readCullen2012}
\title{Read data from Cullen et al. (2012)}
\usage{
readCullen2012(subtype)
}
\arguments{
\item{subtype}{Subtype of Cullen et al. (2012) data to load. Currently
supported subtypes are "flows" and "giMatrix".}
}
\description{
Read data from supplementary material of Cullen et al. 2012 paper
'Mapping the global flow of steel: from steelmaking to end‐use goods'.
Files were digitalized from pdf to Excel.
}
\author{
Merlin Jo Hosak
}
