% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPauliuk2013.R
\name{readPauliuk2013}
\alias{readPauliuk2013}
\title{TODOMERLIN: document}
\usage{
readPauliuk2013(subtype = "sectorSplits")
}
\arguments{
\item{subtype}{TODOMERLIN: document}
}
\description{
TODOMERLIN: document
}
\author{
Merlin Jo Hosak
}
