% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPlastics_CarbonContent.R
\name{readPlastics_CarbonContent}
\alias{readPlastics_CarbonContent}
\title{Read Plastics Carbon Content from stochiometric calculations for defined polymers and
rough estimates for broader classes}
\usage{
readPlastics_CarbonContent()
}
\value{
magpie object of the Carbon Content for different plastic types/polymers
}
\description{
Read Plastics Carbon Content from stochiometric calculations for defined polymers and
rough estimates for broader classes
}
\examples{
\dontrun{
a <- readSource(type = "Plastics_CarbonContent")
}

}
\seealso{
[readSource()]
}
\author{
Leonie Schweiger
}
