% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWorldSteelParameters.R
\name{readWorldSteelParameters}
\alias{readWorldSteelParameters}
\title{Read World Steel Association Parameters}
\usage{
readWorldSteelParameters(subtype)
}
\arguments{
\item{subtype}{TODOMERLIN: document}
}
\description{
Read World Steel Association (WSA) parameters from various sources like
yearbooks, pdfs and webpages. The data is stored in Excel files in the
WorldSteelParameters directory with metadata, the primary sources
are stored there as well if available.
}
\author{
Merlin Jo Hosak
}
