% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGEM.R
\name{toolAggregateFunctionType}
\alias{toolAggregateFunctionType}
\title{Aggregate building strucutre types to match RASMI definition.}
\usage{
toolAggregateFunctionType(structure)
}
\arguments{
\item{structure}{Vector of building structure types as used in GEM.}
}
\description{
Aggregate building strucutre types to match RASMI definition.
}
\author{
Bennet Weiss
}
