% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolInterpolate.R
\name{toolInterpolate}
\alias{toolInterpolate}
\title{Interpolate (and/or extrapolate) 2D magclass objects.}
\usage{
toolInterpolate(
  x,
  interpolate = TRUE,
  extrapolate = FALSE,
  from = NA,
  to = NA,
  method = "linear",
  extrapolate_method = "constant",
  ref = NA
)
}
\arguments{
\item{x}{magpie object with years as columns that has to be interpolated 
and/or extrapolated}

\item{interpolate}{If TRUE, the function will interpolate missing values
(default: TRUE).}

\item{extrapolate}{If TRUE, the function will extrapolate missing values
(default: FALSE).}

\item{from}{Start year for interpolation/extrapolation (default: NA,
which will be set to the first year in x).}

\item{to}{End year for interpolation/extrapolation (default: NA,
which will be set to the last year in sx).}

\item{method}{Interpolation method to use (default: 'linear'). Options:
'linear', 'constant'. # TODO add others}

\item{extrapolate_method}{Extrapolation method to use (default: 'constant').
Options: 'constant', 'ref'. # TODO add more methods. If 'ref', a reference 
magpie object must be provided. In 'ref' extrapolation, the values are 
extrapolated based on the shape of the reference data, see 
\link[=toolExtrapolateRegion]{toolExtrapolateRegion}.}

\item{ref}{Reference magpie object for extrapolation (default: NA).}
}
\description{
Interpolate (and/or extrapolate) 2D magclass objects.
}
\author{
Merlin Jo Hosak
}
