% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolInterpolate.R
\name{toolInterpolate}
\alias{toolInterpolate}
\title{Interpolate missing values in a magpie object.}
\usage{
toolInterpolate(x, years = NULL, type = "linear", extrapolate = FALSE, ...)
}
\arguments{
\item{x}{A magpie object (2D or 3D) with missing values to interpolate.}

\item{years}{Optional vector of years to interpolate to. If NULL, uses years in x.}

\item{type}{Interpolation method: one of 'linear', 'spline', 'monotone', or 'constant'. Default is 'linear'.}

\item{extrapolate}{Logical. Whether to allow extrapolation outside the range of available data. Default is FALSE.}

\item{...}{Additional arguments passed to interpolation functions.}
}
\value{
A magpie object with interpolated values for the specified years and all data slices.
}
\description{
Interpolates missing values (NAs) in a magpie object along the time dimension (years), or to given years.
}
\author{
Bennet Weiss
}
