% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolInterpolate.R
\name{toolInterpolateRegion}
\alias{toolInterpolateRegion}
\title{Interpolate data for a single region}
\usage{
toolInterpolateRegion(row, method = "linear")
}
\description{
Interpolate data for a single region
}
